@@ -16,8 +16,8 @@ namespace common::errormsg::cor {
         ADD(0x80004003, "COR_E_NULLREFERENCE", "Dereferencing a null reference. In general class libraries should not throw this", "corerror.h");
         ADD(0x8002000e, "COR_E_TARGETPARAMCOUNT", "There was a mismatch between number of arguments provided and the number expected", "corerror.h");
         ADD(0x80020012, "COR_E_DIVIDEBYZERO", "Attempted to divide a number by zero.", "corerror.h");
-        ADD(0x8004dead, "E_SYM_DESTROYED", nullptr, "corsym.h");
-        ADD(0x80070002, "COR_E_FILENOTFOUND", nullptr, "corerror.h");
+        ADD(0x8004dead, "E_SYM_DESTROYED", nil, "corsym.h");
+        ADD(0x80070002, "COR_E_FILENOTFOUND", nil, "corerror.h");
         ADD(0x80070003, "COR_E_DIRECTORYNOTFOUND", "The specified path couldn't be found.", "corerror.h");
         ADD(0x80070005, "COR_E_UNAUTHORIZEDACCESS", "Access is denied.", "corerror.h");
         ADD(0x8007000b, "COR_E_BADIMAGEFORMAT", "The format of DLL or executable being loaded is invalid.", "corerror.h");
@@ -33,13 +33,13 @@ namespace common::errormsg::cor {
         ADD(0x80131011, "MSEE_E_GETPROCFAILED", "Failed to get entry point %s (Win32 error: %d).", "corerror.h");
         ADD(0x80131012, "MSEE_E_MULTCOPIESLOADED", "Multiple copies of MSCOREE.dll have been loaded by the same process.", "corerror.h");
         ADD(0x80131013, "COR_E_TYPEUNLOADED", "The type had been unloaded.", "corerror.h");
-        ADD(0x80131013, "COR_E_TYPEUNLOADED", nullptr, "corerror.h");
+        ADD(0x80131013, "COR_E_TYPEUNLOADED", nil, "corerror.h");
         ADD(0x80131014, "COR_E_APPDOMAINUNLOADED", "access unloaded appdomain", "corerror.h");
         ADD(0x80131015, "COR_E_CANNOTUNLOADAPPDOMAIN", "Error while unloading an appdomain", "corerror.h");
         ADD(0x80131016, "MSEE_E_ASSEMBLYLOADINPROGRESS", "Assembly is being currently being loaded", "corerror.h");
         ADD(0x80131017, "MSEE_E_CANNOTCREATEAPPDOMAIN", "Attempt to create appdomain failed", "corerror.h");
         ADD(0x80131018, "COR_E_ASSEMBLYEXPECTED", "The module was expected to contain an assembly manifest.", "corerror.h");
-        ADD(0x80131018, "COR_E_ASSEMBLYEXPECTED", nullptr, "corerror.h");
+        ADD(0x80131018, "COR_E_ASSEMBLYEXPECTED", nil, "corerror.h");
         ADD(0x80131019, "COR_E_FIXUPSINEXE", "Attempt to load an unverifiable exe with fixups (IAT with more than 2 sections or a TLS section)", "corerror.h");
         ADD(0x8013101a, "COR_E_NO_LOADLIBRARY_ALLOWED", "Attempt to LoadLibrary a managed image in an improper way (only assemblies with EAT's area allowed.)", "corerror.h");
         ADD(0x8013101b, "COR_E_NEWER_RUNTIME", "The assembly is built by a runtime newer than the currently loaded runtime, and cannot be loaded.", "corerror.h");
@@ -50,7 +50,7 @@ namespace common::errormsg::cor {
         ADD(0x80131022, "HOST_E_INVALIDOPERATION", "The operation is invalid", "corerror.h");
         ADD(0x80131023, "HOST_E_CLRNOTAVAILABLE", "CLR has been disabled due to unrecoverable error", "corerror.h");
         ADD(0x80131024, "HOST_E_TIMEOUT", "A wait times out", "corerror.h");
-        ADD(0x80131025, "HOST_E_NOT_OWNER", nullptr, "corerror.h");
+        ADD(0x80131025, "HOST_E_NOT_OWNER", nil, "corerror.h");
         ADD(0x80131026, "HOST_E_ABANDONED", "An event is abandoned", "corerror.h");
         ADD(0x80131027, "HOST_E_EXITPROCESS_THREADABORT", "ExitProcess due to ThreadAbort escalation", "corerror.h");
         ADD(0x80131028, "HOST_E_EXITPROCESS_ADUNLOAD", "ExitProcess due to AD Unload escalation", "corerror.h");
@@ -109,7 +109,7 @@ namespace common::errormsg::cor {
         ADD(0x80131161, "TLBX_E_INVALID_TYPEINFO_UNNAMED", "Typelib import: invalid type, not converted -- name unknown.", "corerror.h");
         ADD(0x80131162, "TLBX_E_CTX_NESTED", "Typelib export: Format string for nested contexts.", "corerror.h");
         ADD(0x80131163, "TLBX_E_ERROR_MESSAGE", "Typelib export: Error message wrapper.", "corerror.h");
-        ADD(0x80131164, "TLBX_E_CANT_SAVE", "Typelib export: cant nullptrSaveAllChanges()nullptr", "corerror.h");
+        ADD(0x80131164, "TLBX_E_CANT_SAVE", "Typelib export: cant \"SaveAllChanges()\"", "corerror.h");
         ADD(0x80131165, "TLBX_W_LIBNOTREGISTERED", "Typelib export: type library is not registered.", "corerror.h");
         ADD(0x80131166, "TLBX_E_CANTLOADLIBRARY", "Typelib export: type library cannot be loaded.", "corerror.h");
         ADD(0x80131167, "TLBX_E_BAD_VT_TYPE", "Typelib import: invalid VT_*, not converted.", "corerror.h");
@@ -121,7 +121,7 @@ namespace common::errormsg::cor {
         ADD(0x8013116e, "TLBX_E_DUPLICATE_TYPE_NAME", "Typelib import: duplicate name (due to user-defined name).", "corerror.h");
         ADD(0x80131172, "TLBX_I_NONSEQUENTIALSTRUCT", "Typelib export: Can't convert non-sequential structs.", "corerror.h");
         ADD(0x80131174, "TLBX_I_RESOLVEREFFAILED", "Typelib import: The resolve ref call failed.", "corerror.h");
-        ADD(0x80131175, "TLBX_E_ASANY", "Typelib export: Encounterd nullptrAsAnynullptr -- ignored.", "corerror.h");
+        ADD(0x80131175, "TLBX_E_ASANY", "Typelib export: Encounterd \"AsAny\" -- ignored.", "corerror.h");
         ADD(0x80131176, "TLBX_E_INVALIDLCIDPARAM", "Typelib export: Encounterd an LCID attribute set to an invalid param.", "corerror.h");
         ADD(0x80131177, "TLBX_E_LCIDONDISPONLYITF", "Typelib export: Encounterd an LCID attribute on a disp only interface.", "corerror.h");
         ADD(0x80131178, "TLBX_E_NONPUBLIC_FIELD", "Typelib export: Non-public field in public struct.", "corerror.h");
@@ -312,16 +312,16 @@ namespace common::errormsg::cor {
         ADD(0x8013125e, "VLDTR_E_CT_DUP", "ExportedType has dup.", "corerror.h");
         ADD(0x8013125f, "VLDTR_E_MAR_DUP", "ManifestResource has dup.", "corerror.h");
         ADD(0x80131260, "VLDTR_E_MAR_NOTPUBPRIV", "ManifestResource is neither Public not Private.", "corerror.h");
-        ADD(0x80131261, "VLDTR_E_TD_ENUMNOVALUE", "Enum has no nullptrvalue__nullptr field.", "corerror.h");
-        ADD(0x80131262, "VLDTR_E_TD_ENUMVALSTATIC", "Enum's nullptrvalue__nullptr field is static.", "corerror.h");
-        ADD(0x80131263, "VLDTR_E_TD_ENUMVALNOTSN", "Enum's nullptrvalue__nullptr field is not SpecialName.", "corerror.h");
+        ADD(0x80131261, "VLDTR_E_TD_ENUMNOVALUE", "Enum has no \"value__\" field.", "corerror.h");
+        ADD(0x80131262, "VLDTR_E_TD_ENUMVALSTATIC", "Enum's \"value__\" field is static.", "corerror.h");
+        ADD(0x80131263, "VLDTR_E_TD_ENUMVALNOTSN", "Enum's \"value__\" field is not SpecialName.", "corerror.h");
         ADD(0x80131264, "VLDTR_E_TD_ENUMFLDNOTST", "Enum's field is not static.", "corerror.h");
         ADD(0x80131265, "VLDTR_E_TD_ENUMFLDNOTLIT", "Enum's field is not literal.", "corerror.h");
         ADD(0x80131266, "VLDTR_E_TD_ENUMNOLITFLDS", "Enum has no literal fields.", "corerror.h");
         ADD(0x80131267, "VLDTR_E_TD_ENUMFLDSIGMISMATCH", "Enum's field sig does not match value__ sig.", "corerror.h");
-        ADD(0x80131268, "VLDTR_E_TD_ENUMVALNOT1ST", "Enum's nullptrvalue__nullptr field is not first.", "corerror.h");
-        ADD(0x80131269, "VLDTR_E_FD_NOTVALUERTSN", "Field is RTSpecialName but name is not nullptrvalue__nullptr.", "corerror.h");
-        ADD(0x8013126a, "VLDTR_E_FD_VALUEPARNOTENUM", "Field nullptrvalue__nullptr in not Enum class.", "corerror.h");
+        ADD(0x80131268, "VLDTR_E_TD_ENUMVALNOT1ST", "Enum's \"value__\" field is not first.", "corerror.h");
+        ADD(0x80131269, "VLDTR_E_FD_NOTVALUERTSN", "Field is RTSpecialName but name is not \"value__\".", "corerror.h");
+        ADD(0x8013126a, "VLDTR_E_FD_VALUEPARNOTENUM", "Field \"value__\" in not Enum class.", "corerror.h");
         ADD(0x8013126b, "VLDTR_E_FD_INSTINIFACE", "Instance field in interface.", "corerror.h");
         ADD(0x8013126c, "VLDTR_E_FD_NOTPUBINIFACE", "Non-public field in interface.", "corerror.h");
         ADD(0x8013126d, "VLDTR_E_FMD_GLOBALNOTPUBPRIVSC", "Global field/method neither Public nor PrivateScope.", "corerror.h");
@@ -518,7 +518,7 @@ namespace common::errormsg::cor {
         ADD(0x8013133e, "CORDBG_E_ENC_BAD_METHOD_INFO", "Lacking information about method.", "corerror.h");
         ADD(0x8013133f, "CORDBG_E_ENC_JIT_CANT_UPDATE", "The JIT is unable to update the method.", "corerror.h");
         ADD(0x80131340, "CORDBG_E_ENC_MISSING_CLASS", "An internal structure about the class is missing", "corerror.h");
-        ADD(0x80131341, "CORDBG_E_ENC_INTERNAL_ERROR", "Generic message for nullptrSomething user doesn't control went wrongnullptr message.", "corerror.h");
+        ADD(0x80131341, "CORDBG_E_ENC_INTERNAL_ERROR", "Generic message for \"Something user doesn't control went wrong\" message.", "corerror.h");
         ADD(0x80131342, "CORDBG_E_ENC_HANGING_FIELD", "The field was added via enc after the class was loaded, and so instead of", "corerror.h");
         ADD(0x80131343, "CORDBG_E_MODULE_NOT_LOADED", "If the module isn't loaded, including if it's been unloaded.", "corerror.h");
         ADD(0x80131344, "CORDBG_E_ENC_CANT_CHANGE_SUPERCLASS", "Not allowed to change which class something inherits from", "corerror.h");
@@ -602,417 +602,417 @@ namespace common::errormsg::cor {
         ADD(0x8013144d, "CORSECATTR_E_NO_PROPERTY", "Couldn't find a CA property", "corerror.h");
         ADD(0x8013144e, "CORSECATTR_E_EXCEPTION", "Unexpected exception", "corerror.h");
         ADD(0x8013144f, "CORSECATTR_E_EXCEPTION_HR", "Unexpected exception", "corerror.h");
-        ADD(0x80131450, "ISS_E_ISOSTORE", nullptr, "corerror.h");
-        ADD(0x80131450, "ISS_E_ISOSTORE_START", nullptr, "corerror.h");
-        ADD(0x80131460, "ISS_E_OPEN_STORE_FILE", nullptr, "corerror.h");
-        ADD(0x80131461, "ISS_E_OPEN_FILE_MAPPING", nullptr, "corerror.h");
-        ADD(0x80131462, "ISS_E_MAP_VIEW_OF_FILE", nullptr, "corerror.h");
-        ADD(0x80131463, "ISS_E_GET_FILE_SIZE", nullptr, "corerror.h");
-        ADD(0x80131464, "ISS_E_CREATE_MUTEX", nullptr, "corerror.h");
-        ADD(0x80131465, "ISS_E_LOCK_FAILED", nullptr, "corerror.h");
-        ADD(0x80131466, "ISS_E_FILE_WRITE", nullptr, "corerror.h");
-        ADD(0x80131467, "ISS_E_SET_FILE_POINTER", nullptr, "corerror.h");
-        ADD(0x80131468, "ISS_E_CREATE_DIR", nullptr, "corerror.h");
-        ADD(0x80131469, "ISS_E_STORE_NOT_OPEN", nullptr, "corerror.h");
-        ADD(0x80131480, "ISS_E_CORRUPTED_STORE_FILE", nullptr, "corerror.h");
-        ADD(0x80131481, "ISS_E_STORE_VERSION", nullptr, "corerror.h");
-        ADD(0x80131482, "ISS_E_FILE_NOT_MAPPED", nullptr, "corerror.h");
-        ADD(0x80131483, "ISS_E_BLOCK_SIZE_TOO_SMALL", nullptr, "corerror.h");
-        ADD(0x80131484, "ISS_E_ALLOC_TOO_LARGE", nullptr, "corerror.h");
-        ADD(0x80131485, "ISS_E_USAGE_WILL_EXCEED_QUOTA", nullptr, "corerror.h");
-        ADD(0x80131486, "ISS_E_TABLE_ROW_NOT_FOUND", nullptr, "corerror.h");
-        ADD(0x801314a0, "ISS_E_DEPRECATE", nullptr, "corerror.h");
-        ADD(0x801314a1, "ISS_E_CALLER", nullptr, "corerror.h");
-        ADD(0x801314a2, "ISS_E_PATH_LENGTH", nullptr, "corerror.h");
-        ADD(0x801314a3, "ISS_E_MACHINE", nullptr, "corerror.h");
-        ADD(0x801314a4, "ISS_E_MACHINE_DACL", nullptr, "corerror.h");
-        ADD(0x801314ff, "ISS_E_ISOSTORE_END", nullptr, "corerror.h");
+        ADD(0x80131450, "ISS_E_ISOSTORE", nil, "corerror.h");
+        ADD(0x80131450, "ISS_E_ISOSTORE_START", nil, "corerror.h");
+        ADD(0x80131460, "ISS_E_OPEN_STORE_FILE", nil, "corerror.h");
+        ADD(0x80131461, "ISS_E_OPEN_FILE_MAPPING", nil, "corerror.h");
+        ADD(0x80131462, "ISS_E_MAP_VIEW_OF_FILE", nil, "corerror.h");
+        ADD(0x80131463, "ISS_E_GET_FILE_SIZE", nil, "corerror.h");
+        ADD(0x80131464, "ISS_E_CREATE_MUTEX", nil, "corerror.h");
+        ADD(0x80131465, "ISS_E_LOCK_FAILED", nil, "corerror.h");
+        ADD(0x80131466, "ISS_E_FILE_WRITE", nil, "corerror.h");
+        ADD(0x80131467, "ISS_E_SET_FILE_POINTER", nil, "corerror.h");
+        ADD(0x80131468, "ISS_E_CREATE_DIR", nil, "corerror.h");
+        ADD(0x80131469, "ISS_E_STORE_NOT_OPEN", nil, "corerror.h");
+        ADD(0x80131480, "ISS_E_CORRUPTED_STORE_FILE", nil, "corerror.h");
+        ADD(0x80131481, "ISS_E_STORE_VERSION", nil, "corerror.h");
+        ADD(0x80131482, "ISS_E_FILE_NOT_MAPPED", nil, "corerror.h");
+        ADD(0x80131483, "ISS_E_BLOCK_SIZE_TOO_SMALL", nil, "corerror.h");
+        ADD(0x80131484, "ISS_E_ALLOC_TOO_LARGE", nil, "corerror.h");
+        ADD(0x80131485, "ISS_E_USAGE_WILL_EXCEED_QUOTA", nil, "corerror.h");
+        ADD(0x80131486, "ISS_E_TABLE_ROW_NOT_FOUND", nil, "corerror.h");
+        ADD(0x801314a0, "ISS_E_DEPRECATE", nil, "corerror.h");
+        ADD(0x801314a1, "ISS_E_CALLER", nil, "corerror.h");
+        ADD(0x801314a2, "ISS_E_PATH_LENGTH", nil, "corerror.h");
+        ADD(0x801314a3, "ISS_E_MACHINE", nil, "corerror.h");
+        ADD(0x801314a4, "ISS_E_MACHINE_DACL", nil, "corerror.h");
+        ADD(0x801314ff, "ISS_E_ISOSTORE_END", nil, "corerror.h");
         ADD(0x80131500, "COR_E_EXCEPTION", "Base class for all exceptions in the runtime", "corerror.h");
-        ADD(0x80131500, "COR_E_EXCEPTION", nullptr, "corerror.h");
-        ADD(0x80131501, "COR_E_SYSTEM", "The base class for the runtime's nullptrless seriousnullptr exceptions", "corerror.h");
-        ADD(0x80131501, "COR_E_SYSTEM", nullptr, "corerror.h");
+        ADD(0x80131500, "COR_E_EXCEPTION", nil, "corerror.h");
+        ADD(0x80131501, "COR_E_SYSTEM", "The base class for the runtime's \"less serious\" exceptions", "corerror.h");
+        ADD(0x80131501, "COR_E_SYSTEM", nil, "corerror.h");
         ADD(0x80131502, "COR_E_ARGUMENTOUTOFRANGE", "An argument was out of its legal range.", "corerror.h");
-        ADD(0x80131502, "COR_E_ARGUMENTOUTOFRANGE", nullptr, "corerror.h");
+        ADD(0x80131502, "COR_E_ARGUMENTOUTOFRANGE", nil, "corerror.h");
         ADD(0x80131503, "COR_E_ARRAYTYPEMISMATCH", "Attempted to store an object of the wrong type in an array", "corerror.h");
-        ADD(0x80131503, "COR_E_ARRAYTYPEMISMATCH", nullptr, "corerror.h");
-        ADD(0x80131504, "COR_E_CONTEXTMARSHAL", nullptr, "corerror.h");
-        ADD(0x80131504, "COR_E_CONTEXTMARSHAL", nullptr, "corerror.h");
-        ADD(0x80131505, "COR_E_TIMEOUT", nullptr, "corerror.h");
-        ADD(0x80131505, "COR_E_TIMEOUT", nullptr, "corerror.h");
+        ADD(0x80131503, "COR_E_ARRAYTYPEMISMATCH", nil, "corerror.h");
+        ADD(0x80131504, "COR_E_CONTEXTMARSHAL", nil, "corerror.h");
+        ADD(0x80131504, "COR_E_CONTEXTMARSHAL", nil, "corerror.h");
+        ADD(0x80131505, "COR_E_TIMEOUT", nil, "corerror.h");
+        ADD(0x80131505, "COR_E_TIMEOUT", nil, "corerror.h");
         ADD(0x80131506, "COR_E_EXECUTIONENGINE", "An internal error happened in the Common Language Runtime's Execution Engine", "corerror.h");
-        ADD(0x80131506, "COR_E_EXECUTIONENGINE", nullptr, "corerror.h");
+        ADD(0x80131506, "COR_E_EXECUTIONENGINE", nil, "corerror.h");
         ADD(0x80131507, "COR_E_FIELDACCESS", "Access to this field is denied.", "corerror.h");
-        ADD(0x80131507, "COR_E_FIELDACCESS", nullptr, "corerror.h");
+        ADD(0x80131507, "COR_E_FIELDACCESS", nil, "corerror.h");
         ADD(0x80131508, "COR_E_INDEXOUTOFRANGE", "Attempted to access an element within an array by using an index that is not within the bound of that array.", "corerror.h");
-        ADD(0x80131508, "COR_E_INDEXOUTOFRANGE", nullptr, "corerror.h");
+        ADD(0x80131508, "COR_E_INDEXOUTOFRANGE", nil, "corerror.h");
         ADD(0x80131509, "COR_E_INVALIDOPERATION", "An operation is not legal in the current state.", "corerror.h");
-        ADD(0x80131509, "COR_E_INVALIDOPERATION", nullptr, "corerror.h");
+        ADD(0x80131509, "COR_E_INVALIDOPERATION", nil, "corerror.h");
         ADD(0x8013150a, "COR_E_SECURITY", "An error relating to security occured.", "corerror.h");
-        ADD(0x8013150a, "COR_E_SECURITY", nullptr, "corerror.h");
+        ADD(0x8013150a, "COR_E_SECURITY", nil, "corerror.h");
         ADD(0x8013150b, "COR_E_REMOTING", "An error relating to remoting occurred.", "corerror.h");
-        ADD(0x8013150b, "COR_E_REMOTING", nullptr, "corerror.h");
-        ADD(0x8013150c, "COR_E_SERIALIZATION", nullptr, "corerror.h");
+        ADD(0x8013150b, "COR_E_REMOTING", nil, "corerror.h");
+        ADD(0x8013150c, "COR_E_SERIALIZATION", nil, "corerror.h");
         ADD(0x8013150d, "COR_E_VERIFICATION", "A verification failure occurred", "corerror.h");
-        ADD(0x8013150d, "COR_E_VERIFICATION", nullptr, "corerror.h");
-        ADD(0x8013150e, "COR_E_SERVER", nullptr, "corerror.h");
+        ADD(0x8013150d, "COR_E_VERIFICATION", nil, "corerror.h");
+        ADD(0x8013150e, "COR_E_SERVER", nil, "corerror.h");
         ADD(0x8013150f, "COR_E_SERVICEDCOMPONENT", "An error relating to ServicedComponent occurred.", "corerror.h");
-        ADD(0x8013150f, "COR_E_SERVICEDCOMPONENT", nullptr, "corerror.h");
+        ADD(0x8013150f, "COR_E_SERVICEDCOMPONENT", nil, "corerror.h");
         ADD(0x80131510, "COR_E_METHODACCESS", "Access to this method is denied.", "corerror.h");
-        ADD(0x80131510, "COR_E_METHODACCESS", nullptr, "corerror.h");
+        ADD(0x80131510, "COR_E_METHODACCESS", nil, "corerror.h");
         ADD(0x80131511, "COR_E_MISSINGFIELD", "An attempt was made to dynamically access a field that does not exist.", "corerror.h");
-        ADD(0x80131511, "COR_E_MISSINGFIELD", nullptr, "corerror.h");
+        ADD(0x80131511, "COR_E_MISSINGFIELD", nil, "corerror.h");
         ADD(0x80131512, "COR_E_MISSINGMEMBER", "An attempt was made to dynamically invoke or access a field or method that does not exist.", "corerror.h");
-        ADD(0x80131512, "COR_E_MISSINGMEMBER", nullptr, "corerror.h");
+        ADD(0x80131512, "COR_E_MISSINGMEMBER", nil, "corerror.h");
         ADD(0x80131513, "COR_E_MISSINGMETHOD", "An attempt was made to dynamically invoke a method that does not exist", "corerror.h");
-        ADD(0x80131513, "COR_E_MISSINGMETHOD", nullptr, "corerror.h");
+        ADD(0x80131513, "COR_E_MISSINGMETHOD", nil, "corerror.h");
         ADD(0x80131514, "COR_E_MULTICASTNOTSUPPORTED", "Attempted to combine delegates that are not multicast", "corerror.h");
-        ADD(0x80131514, "COR_E_MULTICASTNOTSUPPORTED", nullptr, "corerror.h");
+        ADD(0x80131514, "COR_E_MULTICASTNOTSUPPORTED", nil, "corerror.h");
         ADD(0x80131515, "COR_E_NOTSUPPORTED", "The operation is not supported", "corerror.h");
-        ADD(0x80131515, "COR_E_NOTSUPPORTED", nullptr, "corerror.h");
+        ADD(0x80131515, "COR_E_NOTSUPPORTED", nil, "corerror.h");
         ADD(0x80131516, "COR_E_OVERFLOW", "An arithmetic, casting, or conversion operation overflowed or underflowed.", "corerror.h");
-        ADD(0x80131516, "COR_E_OVERFLOW", nullptr, "corerror.h");
+        ADD(0x80131516, "COR_E_OVERFLOW", nil, "corerror.h");
         ADD(0x80131517, "COR_E_RANK", "An array has the wrong number of dimensions for a particular operation.", "corerror.h");
-        ADD(0x80131517, "COR_E_RANK", nullptr, "corerror.h");
+        ADD(0x80131517, "COR_E_RANK", nil, "corerror.h");
         ADD(0x80131518, "COR_E_SYNCHRONIZATIONLOCK", "Wait(), Notify() or NotifyAll() was called from an unsynchronized ** block of c", "corerror.h");
-        ADD(0x80131518, "COR_E_SYNCHRONIZATIONLOCK", nullptr, "corerror.h");
+        ADD(0x80131518, "COR_E_SYNCHRONIZATIONLOCK", nil, "corerror.h");
         ADD(0x80131519, "COR_E_THREADINTERRUPTED", "Indicates that the thread was interrupted from a waiting state", "corerror.h");
-        ADD(0x80131519, "COR_E_THREADINTERRUPTED", nullptr, "corerror.h");
+        ADD(0x80131519, "COR_E_THREADINTERRUPTED", nil, "corerror.h");
         ADD(0x8013151a, "COR_E_MEMBERACCESS", "Access to this member is denied.", "corerror.h");
-        ADD(0x8013151a, "COR_E_MEMBERACCESS", nullptr, "corerror.h");
+        ADD(0x8013151a, "COR_E_MEMBERACCESS", nil, "corerror.h");
         ADD(0x80131520, "COR_E_THREADSTATE", "Indicate that the Thread class is in an invalid state for the method call", "corerror.h");
-        ADD(0x80131520, "COR_E_THREADSTATE", nullptr, "corerror.h");
+        ADD(0x80131520, "COR_E_THREADSTATE", nil, "corerror.h");
         ADD(0x80131521, "COR_E_THREADSTOP", "Thrown into a thread to cause it to stop. This exception is typically not caught", "corerror.h");
-        ADD(0x80131521, "COR_E_THREADSTOP", nullptr, "corerror.h");
+        ADD(0x80131521, "COR_E_THREADSTOP", nil, "corerror.h");
         ADD(0x80131522, "COR_E_TYPELOAD", "Could not find or load a specific type (class, enum, etc).", "corerror.h");
-        ADD(0x80131522, "COR_E_TYPELOAD", nullptr, "corerror.h");
+        ADD(0x80131522, "COR_E_TYPELOAD", nil, "corerror.h");
         ADD(0x80131523, "COR_E_ENTRYPOINTNOTFOUND", "Could not find the specified DllImport entry point", "corerror.h");
-        ADD(0x80131523, "COR_E_ENTRYPOINTNOTFOUND", nullptr, "corerror.h");
+        ADD(0x80131523, "COR_E_ENTRYPOINTNOTFOUND", nil, "corerror.h");
         ADD(0x80131524, "COR_E_DLLNOTFOUND", "Could not find the specified DllImport DLL.", "corerror.h");
-        ADD(0x80131524, "COR_E_DLLNOTFOUND", nullptr, "corerror.h");
+        ADD(0x80131524, "COR_E_DLLNOTFOUND", nil, "corerror.h");
         ADD(0x80131525, "COR_E_THREADSTART", "Indicate that a user thread fails to start.", "corerror.h");
-        ADD(0x80131525, "COR_E_THREADSTART", nullptr, "corerror.h");
+        ADD(0x80131525, "COR_E_THREADSTART", nil, "corerror.h");
         ADD(0x80131527, "COR_E_INVALIDCOMOBJECT", "An invalid __ComObject has been used.", "corerror.h");
-        ADD(0x80131527, "COR_E_INVALIDCOMOBJECT", nullptr, "corerror.h");
+        ADD(0x80131527, "COR_E_INVALIDCOMOBJECT", nil, "corerror.h");
         ADD(0x80131528, "COR_E_NOTFINITENUMBER", "Thrown if value (a floating point number) is either the not a number value (NaN) or +- infinity value VB needs this stuff", "corerror.h");
-        ADD(0x80131528, "COR_E_NOTFINITENUMBER", nullptr, "corerror.h");
+        ADD(0x80131528, "COR_E_NOTFINITENUMBER", nil, "corerror.h");
         ADD(0x80131529, "COR_E_DUPLICATEWAITOBJECT", "An object appears more than once in the wait objects array.", "corerror.h");
-        ADD(0x80131529, "COR_E_DUPLICATEWAITOBJECT", nullptr, "corerror.h");
+        ADD(0x80131529, "COR_E_DUPLICATEWAITOBJECT", nil, "corerror.h");
         ADD(0x8013152b, "COR_E_SEMAPHOREFULL", "Adding the given count to the semaphore would cause it to exceed its maximum count.", "corerror.h");
-        ADD(0x8013152b, "COR_E_SEMAPHOREFULL", nullptr, "corerror.h");
+        ADD(0x8013152b, "COR_E_SEMAPHOREFULL", nil, "corerror.h");
         ADD(0x8013152c, "COR_E_WAITHANDLECANNOTBEOPENED", "No Semaphore of the given name exists.", "corerror.h");
-        ADD(0x8013152c, "COR_E_WAITHANDLECANNOTBEOPENED", nullptr, "corerror.h");
+        ADD(0x8013152c, "COR_E_WAITHANDLECANNOTBEOPENED", nil, "corerror.h");
         ADD(0x8013152d, "COR_E_ABANDONEDMUTEX", "The wait completed due to an abandoned mutex.", "corerror.h");
-        ADD(0x8013152d, "COR_E_ABANDONEDMUTEX", nullptr, "corerror.h");
+        ADD(0x8013152d, "COR_E_ABANDONEDMUTEX", nil, "corerror.h");
         ADD(0x80131530, "COR_E_THREADABORTED", "Thrown into a thread to cause it to abort. Not catchable.", "corerror.h");
-        ADD(0x80131530, "COR_E_THREADABORTED", nullptr, "corerror.h");
+        ADD(0x80131530, "COR_E_THREADABORTED", nil, "corerror.h");
         ADD(0x80131531, "COR_E_INVALIDOLEVARIANTTYPE", "The type of an OLE variant that was passed into the runtime is invalid.", "corerror.h");
-        ADD(0x80131531, "COR_E_INVALIDOLEVARIANTTYPE", nullptr, "corerror.h");
+        ADD(0x80131531, "COR_E_INVALIDOLEVARIANTTYPE", nil, "corerror.h");
         ADD(0x80131532, "COR_E_MISSINGMANIFESTRESOURCE", "An expected resource in the assembly manifest was missing.", "corerror.h");
-        ADD(0x80131532, "COR_E_MISSINGMANIFESTRESOURCE", nullptr, "corerror.h");
+        ADD(0x80131532, "COR_E_MISSINGMANIFESTRESOURCE", nil, "corerror.h");
         ADD(0x80131533, "COR_E_SAFEARRAYTYPEMISMATCH", "A mismatch has occured between the runtime type of the array and the sub type recorded in the metadata.", "corerror.h");
-        ADD(0x80131533, "COR_E_SAFEARRAYTYPEMISMATCH", nullptr, "corerror.h");
+        ADD(0x80131533, "COR_E_SAFEARRAYTYPEMISMATCH", nil, "corerror.h");
         ADD(0x80131534, "COR_E_TYPEINITIALIZATION", "An exception was thrown by a type's initializer (.cctor).", "corerror.h");
-        ADD(0x80131534, "COR_E_TYPEINITIALIZATION", nullptr, "corerror.h");
+        ADD(0x80131534, "COR_E_TYPEINITIALIZATION", nil, "corerror.h");
         ADD(0x80131535, "COR_E_MARSHALDIRECTIVE", "The marshaling directives are invalid.", "corerror.h");
-        ADD(0x80131535, "COR_E_MARSHALDIRECTIVE", nullptr, "corerror.h");
+        ADD(0x80131535, "COR_E_MARSHALDIRECTIVE", nil, "corerror.h");
         ADD(0x80131536, "COR_E_MISSINGSATELLITEASSEMBLY", "An expected satellite assembly containing the ultimate fallback resources for a given culture was not found or couldn't be loaded.  Setup problem?", "corerror.h");
-        ADD(0x80131536, "COR_E_MISSINGSATELLITEASSEMBLY", nullptr, "corerror.h");
+        ADD(0x80131536, "COR_E_MISSINGSATELLITEASSEMBLY", nil, "corerror.h");
         ADD(0x80131537, "COR_E_FORMAT", "The format of one arguments does not meet the contract of the method.", "corerror.h");
-        ADD(0x80131537, "COR_E_FORMAT", nullptr, "corerror.h");
+        ADD(0x80131537, "COR_E_FORMAT", nil, "corerror.h");
         ADD(0x80131538, "COR_E_SAFEARRAYRANKMISMATCH", "A mismatch has occured between the runtime rank of the array and the rank recorded in the metadata.", "corerror.h");
-        ADD(0x80131538, "COR_E_SAFEARRAYRANKMISMATCH", nullptr, "corerror.h");
+        ADD(0x80131538, "COR_E_SAFEARRAYRANKMISMATCH", nil, "corerror.h");
         ADD(0x80131539, "COR_E_PLATFORMNOTSUPPORTED", "The method is not supported on this platform", "corerror.h");
-        ADD(0x80131539, "COR_E_PLATFORMNOTSUPPORTED", nullptr, "corerror.h");
+        ADD(0x80131539, "COR_E_PLATFORMNOTSUPPORTED", nil, "corerror.h");
         ADD(0x8013153a, "COR_E_INVALIDPROGRAM", "A program contained invalid IL or bad metadata.  Usually this is a compiler bug.", "corerror.h");
-        ADD(0x8013153a, "COR_E_INVALIDPROGRAM", nullptr, "corerror.h");
+        ADD(0x8013153a, "COR_E_INVALIDPROGRAM", nil, "corerror.h");
         ADD(0x8013153b, "COR_E_OPERATIONCANCELED", "The operation was cancelled.", "corerror.h");
-        ADD(0x8013153b, "COR_E_OPERATIONCANCELED", nullptr, "corerror.h");
+        ADD(0x8013153b, "COR_E_OPERATIONCANCELED", nil, "corerror.h");
         ADD(0x8013153d, "COR_E_INSUFFICIENTMEMORY", "Not enough memory was available for an operation. This may not be potentially fatal (vs. an OutOfMemoryException).", "corerror.h");
-        ADD(0x8013153d, "COR_E_INSUFFICIENTMEMORY", nullptr, "corerror.h");
+        ADD(0x8013153d, "COR_E_INSUFFICIENTMEMORY", nil, "corerror.h");
         ADD(0x8013153e, "COR_E_RUNTIMEWRAPPED", "An object that does not derive from System.Exception has been wrapped in a RuntimeWrappedException.", "corerror.h");
-        ADD(0x8013153e, "COR_E_RUNTIMEWRAPPED", nullptr, "corerror.h");
-        ADD(0x80131540, "COR_E_DEVICESNOTSUPPORTED", nullptr, "corerror.h");
-        ADD(0x80131540, "COR_E_DEVICESNOTSUPPORTED", nullptr, "corerror.h");
+        ADD(0x8013153e, "COR_E_RUNTIMEWRAPPED", nil, "corerror.h");
+        ADD(0x80131540, "COR_E_DEVICESNOTSUPPORTED", nil, "corerror.h");
+        ADD(0x80131540, "COR_E_DEVICESNOTSUPPORTED", nil, "corerror.h");
         ADD(0x80131541, "COR_E_DATAMISALIGNED", "A datatype misalignment was detected in a load or store instruction.", "corerror.h");
-        ADD(0x80131541, "COR_E_DATAMISALIGNED", nullptr, "corerror.h");
-        ADD(0x80131577, "COR_E_KEYNOTFOUND", nullptr, "corerror.h");
-        ADD(0x80131577, "COR_E_KEYNOTFOUND", nullptr, "corerror.h");
-        ADD(0x80131600, "COR_E_APPLICATION", "The base class for all nullptrless seriousnullptr exceptions.", "corerror.h");
-        ADD(0x80131600, "COR_E_APPLICATION", nullptr, "corerror.h");
+        ADD(0x80131541, "COR_E_DATAMISALIGNED", nil, "corerror.h");
+        ADD(0x80131577, "COR_E_KEYNOTFOUND", nil, "corerror.h");
+        ADD(0x80131577, "COR_E_KEYNOTFOUND", nil, "corerror.h");
+        ADD(0x80131600, "COR_E_APPLICATION", "The base class for all \"less serious\" exceptions.", "corerror.h");
+        ADD(0x80131600, "COR_E_APPLICATION", nil, "corerror.h");
         ADD(0x80131601, "COR_E_INVALIDFILTERCRITERIA", "The given filter criteria does not match the filter contract.", "corerror.h");
-        ADD(0x80131601, "COR_E_INVALIDFILTERCRITERIA", nullptr, "corerror.h");
+        ADD(0x80131601, "COR_E_INVALIDFILTERCRITERIA", nil, "corerror.h");
         ADD(0x80131602, "COR_E_REFLECTIONTYPELOAD", "Could not find or load a specific class that was requested through Reflection", "corerror.h");
-        ADD(0x80131602, "COR_E_REFLECTIONTYPELOAD", nullptr, "corerror.h");
-        ADD(0x80131603, "COR_E_TARGET", nullptr, "corerror.h");
-        ADD(0x80131603, "COR_E_TARGET", nullptr, "corerror.h");
+        ADD(0x80131602, "COR_E_REFLECTIONTYPELOAD", nil, "corerror.h");
+        ADD(0x80131603, "COR_E_TARGET", nil, "corerror.h");
+        ADD(0x80131603, "COR_E_TARGET", nil, "corerror.h");
         ADD(0x80131604, "COR_E_TARGETINVOCATION", "If the method called throws an exception", "corerror.h");
-        ADD(0x80131604, "COR_E_TARGETINVOCATION", nullptr, "corerror.h");
+        ADD(0x80131604, "COR_E_TARGETINVOCATION", nil, "corerror.h");
         ADD(0x80131605, "COR_E_CUSTOMATTRIBUTEFORMAT", "If the binary format of a custom attribute is invalid.", "corerror.h");
-        ADD(0x80131605, "COR_E_CUSTOMATTRIBUTEFORMAT", nullptr, "corerror.h");
+        ADD(0x80131605, "COR_E_CUSTOMATTRIBUTEFORMAT", nil, "corerror.h");
         ADD(0x80131620, "COR_E_IO", "Some sort of I/O error.", "corerror.h");
-        ADD(0x80131620, "COR_E_IO", nullptr, "corerror.h");
-        ADD(0x80131621, "COR_E_FILELOAD", nullptr, "corerror.h");
-        ADD(0x80131621, "COR_E_FILELOAD", nullptr, "corerror.h");
+        ADD(0x80131620, "COR_E_IO", nil, "corerror.h");
+        ADD(0x80131621, "COR_E_FILELOAD", nil, "corerror.h");
+        ADD(0x80131621, "COR_E_FILELOAD", nil, "corerror.h");
         ADD(0x80131622, "COR_E_OBJECTDISPOSED", "The object has already been disposed.", "corerror.h");
-        ADD(0x80131622, "COR_E_OBJECTDISPOSED", nullptr, "corerror.h");
+        ADD(0x80131622, "COR_E_OBJECTDISPOSED", nil, "corerror.h");
         ADD(0x80131623, "COR_E_FAILFAST", "Runtime operation halted by call to System.Environment.FailFast().", "corerror.h");
-        ADD(0x80131623, "COR_E_FAILFAST", nullptr, "corerror.h");
+        ADD(0x80131623, "COR_E_FAILFAST", nil, "corerror.h");
         ADD(0x80131640, "COR_E_HOSTPROTECTION", "Attempted to perform an operation that was forbidden by the host.", "corerror.h");
-        ADD(0x80131640, "COR_E_HOSTPROTECTION", nullptr, "corerror.h");
+        ADD(0x80131640, "COR_E_HOSTPROTECTION", nil, "corerror.h");
         ADD(0x80131641, "COR_E_ILLEGAL_REENTRANCY", "Attempted to call into managed code when executing inside a low level extensibility point.", "corerror.h");
-        ADD(0x80131641, "COR_E_ILLEGAL_REENTRANCY", nullptr, "corerror.h");
+        ADD(0x80131641, "COR_E_ILLEGAL_REENTRANCY", nil, "corerror.h");
         ADD(0x80131700, "CLR_E_SHIM_RUNTIMELOAD", "Failed to load the runtime", "corerror.h");
         ADD(0x80131701, "CLR_E_SHIM_RUNTIMEEXPORT", "Failed to find a required export in the runtime", "corerror.h");
         ADD(0x80131702, "CLR_E_SHIM_INSTALLROOT", "Install root is not defined", "corerror.h");
         ADD(0x80131703, "CLR_E_SHIM_INSTALLCOMP", "Expected component of the runtime is not available", "corerror.h");
-        ADD(0x80131801, "VER_E_HRESULT", nullptr, "corerror.h");
-        ADD(0x80131802, "VER_E_OFFSET", nullptr, "corerror.h");
-        ADD(0x80131803, "VER_E_OPCODE", nullptr, "corerror.h");
-        ADD(0x80131804, "VER_E_OPERAND", nullptr, "corerror.h");
-        ADD(0x80131805, "VER_E_TOKEN", nullptr, "corerror.h");
-        ADD(0x80131806, "VER_E_EXCEPT", nullptr, "corerror.h");
-        ADD(0x80131807, "VER_E_STACK_SLOT", nullptr, "corerror.h");
-        ADD(0x80131808, "VER_E_LOC", nullptr, "corerror.h");
-        ADD(0x80131809, "VER_E_ARG", nullptr, "corerror.h");
-        ADD(0x8013180a, "VER_E_FOUND", nullptr, "corerror.h");
-        ADD(0x8013180b, "VER_E_EXPECTED", nullptr, "corerror.h");
-        ADD(0x8013180c, "VER_E_LOC_BYNAME", nullptr, "corerror.h");
-        ADD(0x80131810, "VER_E_UNKNOWN_OPCODE", nullptr, "corerror.h");
-        ADD(0x80131811, "VER_E_SIG_CALLCONV", nullptr, "corerror.h");
-        ADD(0x80131812, "VER_E_SIG_ELEMTYPE", nullptr, "corerror.h");
-        ADD(0x80131814, "VER_E_RET_SIG", nullptr, "corerror.h");
-        ADD(0x80131815, "VER_E_FIELD_SIG", nullptr, "corerror.h");
-        ADD(0x80131818, "VER_E_INTERNAL", nullptr, "corerror.h");
-        ADD(0x80131819, "VER_E_STACK_TOO_LARGE", nullptr, "corerror.h");
-        ADD(0x8013181a, "VER_E_ARRAY_NAME_LONG", nullptr, "corerror.h");
-        ADD(0x80131820, "VER_E_FALLTHRU", nullptr, "corerror.h");
-        ADD(0x80131821, "VER_E_TRY_GTEQ_END", nullptr, "corerror.h");
-        ADD(0x80131822, "VER_E_TRYEND_GT_CS", nullptr, "corerror.h");
-        ADD(0x80131823, "VER_E_HND_GTEQ_END", nullptr, "corerror.h");
-        ADD(0x80131824, "VER_E_HNDEND_GT_CS", nullptr, "corerror.h");
-        ADD(0x80131825, "VER_E_FLT_GTEQ_CS", nullptr, "corerror.h");
-        ADD(0x80131826, "VER_E_TRY_START", nullptr, "corerror.h");
-        ADD(0x80131827, "VER_E_HND_START", nullptr, "corerror.h");
-        ADD(0x80131828, "VER_E_FLT_START", nullptr, "corerror.h");
-        ADD(0x80131829, "VER_E_TRY_OVERLAP", nullptr, "corerror.h");
-        ADD(0x8013182a, "VER_E_TRY_EQ_HND_FIL", nullptr, "corerror.h");
-        ADD(0x8013182b, "VER_E_TRY_SHARE_FIN_FAL", nullptr, "corerror.h");
-        ADD(0x8013182c, "VER_E_HND_OVERLAP", nullptr, "corerror.h");
-        ADD(0x8013182d, "VER_E_HND_EQ", nullptr, "corerror.h");
-        ADD(0x8013182e, "VER_E_FIL_OVERLAP", nullptr, "corerror.h");
-        ADD(0x8013182f, "VER_E_FIL_EQ", nullptr, "corerror.h");
-        ADD(0x80131830, "VER_E_FIL_CONT_TRY", nullptr, "corerror.h");
-        ADD(0x80131831, "VER_E_FIL_CONT_HND", nullptr, "corerror.h");
-        ADD(0x80131832, "VER_E_FIL_CONT_FIL", nullptr, "corerror.h");
-        ADD(0x80131833, "VER_E_FIL_GTEQ_CS", nullptr, "corerror.h");
-        ADD(0x80131834, "VER_E_FIL_START", nullptr, "corerror.h");
-        ADD(0x80131835, "VER_E_FALLTHRU_EXCEP", nullptr, "corerror.h");
-        ADD(0x80131836, "VER_E_FALLTHRU_INTO_HND", nullptr, "corerror.h");
-        ADD(0x80131837, "VER_E_FALLTHRU_INTO_FIL", nullptr, "corerror.h");
-        ADD(0x80131838, "VER_E_LEAVE", nullptr, "corerror.h");
-        ADD(0x80131839, "VER_E_RETHROW", nullptr, "corerror.h");
-        ADD(0x8013183a, "VER_E_ENDFINALLY", nullptr, "corerror.h");
-        ADD(0x8013183b, "VER_E_ENDFILTER", nullptr, "corerror.h");
-        ADD(0x8013183c, "VER_E_ENDFILTER_MISSING", nullptr, "corerror.h");
-        ADD(0x8013183d, "VER_E_BR_INTO_TRY", nullptr, "corerror.h");
-        ADD(0x8013183e, "VER_E_BR_INTO_HND", nullptr, "corerror.h");
-        ADD(0x8013183f, "VER_E_BR_INTO_FIL", nullptr, "corerror.h");
-        ADD(0x80131840, "VER_E_BR_OUTOF_TRY", nullptr, "corerror.h");
-        ADD(0x80131841, "VER_E_BR_OUTOF_HND", nullptr, "corerror.h");
-        ADD(0x80131842, "VER_E_BR_OUTOF_FIL", nullptr, "corerror.h");
-        ADD(0x80131843, "VER_E_BR_OUTOF_FIN", nullptr, "corerror.h");
-        ADD(0x80131844, "VER_E_RET_FROM_TRY", nullptr, "corerror.h");
-        ADD(0x80131845, "VER_E_RET_FROM_HND", nullptr, "corerror.h");
-        ADD(0x80131846, "VER_E_RET_FROM_FIL", nullptr, "corerror.h");
-        ADD(0x80131847, "VER_E_BAD_JMP_TARGET", nullptr, "corerror.h");
-        ADD(0x80131848, "VER_E_PATH_LOC", nullptr, "corerror.h");
-        ADD(0x80131849, "VER_E_PATH_THIS", nullptr, "corerror.h");
-        ADD(0x8013184a, "VER_E_PATH_STACK", nullptr, "corerror.h");
-        ADD(0x8013184b, "VER_E_PATH_STACK_DEPTH", nullptr, "corerror.h");
-        ADD(0x8013184c, "VER_E_THIS", nullptr, "corerror.h");
-        ADD(0x8013184d, "VER_E_THIS_UNINIT_EXCEP", nullptr, "corerror.h");
-        ADD(0x8013184e, "VER_E_THIS_UNINIT_STORE", nullptr, "corerror.h");
-        ADD(0x8013184f, "VER_E_THIS_UNINIT_RET", nullptr, "corerror.h");
-        ADD(0x80131850, "VER_E_THIS_UNINIT_V_RET", nullptr, "corerror.h");
-        ADD(0x80131851, "VER_E_THIS_UNINIT_BR", nullptr, "corerror.h");
-        ADD(0x80131852, "VER_E_LDFTN_CTOR", nullptr, "corerror.h");
-        ADD(0x80131853, "VER_E_STACK_NOT_EQ", nullptr, "corerror.h");
-        ADD(0x80131854, "VER_E_STACK_UNEXPECTED", nullptr, "corerror.h");
-        ADD(0x80131855, "VER_E_STACK_EXCEPTION", nullptr, "corerror.h");
-        ADD(0x80131856, "VER_E_STACK_OVERFLOW", nullptr, "corerror.h");
-        ADD(0x80131857, "VER_E_STACK_UNDERFLOW", nullptr, "corerror.h");
-        ADD(0x80131858, "VER_E_STACK_EMPTY", nullptr, "corerror.h");
-        ADD(0x80131859, "VER_E_STACK_UNINIT", nullptr, "corerror.h");
-        ADD(0x8013185a, "VER_E_STACK_I_I4_I8", nullptr, "corerror.h");
-        ADD(0x8013185b, "VER_E_STACK_R_R4_R8", nullptr, "corerror.h");
-        ADD(0x8013185c, "VER_E_STACK_NO_R_I8", nullptr, "corerror.h");
-        ADD(0x8013185d, "VER_E_STACK_NUMERIC", nullptr, "corerror.h");
-        ADD(0x8013185e, "VER_E_STACK_OBJREF", nullptr, "corerror.h");
-        ADD(0x8013185f, "VER_E_STACK_P_OBJREF", nullptr, "corerror.h");
-        ADD(0x80131860, "VER_E_STACK_BYREF", nullptr, "corerror.h");
-        ADD(0x80131861, "VER_E_STACK_METHOD", nullptr, "corerror.h");
-        ADD(0x80131862, "VER_E_STACK_ARRAY_SD", nullptr, "corerror.h");
-        ADD(0x80131863, "VER_E_STACK_VALCLASS", nullptr, "corerror.h");
-        ADD(0x80131864, "VER_E_STACK_P_VALCLASS", nullptr, "corerror.h");
-        ADD(0x80131865, "VER_E_STACK_NO_VALCLASS", nullptr, "corerror.h");
-        ADD(0x80131866, "VER_E_LOC_DEAD", nullptr, "corerror.h");
-        ADD(0x80131867, "VER_E_LOC_NUM", nullptr, "corerror.h");
-        ADD(0x80131868, "VER_E_ARG_NUM", nullptr, "corerror.h");
-        ADD(0x80131869, "VER_E_TOKEN_RESOLVE", nullptr, "corerror.h");
-        ADD(0x8013186a, "VER_E_TOKEN_TYPE", nullptr, "corerror.h");
-        ADD(0x8013186b, "VER_E_TOKEN_TYPE_MEMBER", nullptr, "corerror.h");
-        ADD(0x8013186c, "VER_E_TOKEN_TYPE_FIELD", nullptr, "corerror.h");
-        ADD(0x8013186d, "VER_E_TOKEN_TYPE_SIG", nullptr, "corerror.h");
-        ADD(0x8013186e, "VER_E_UNVERIFIABLE", nullptr, "corerror.h");
-        ADD(0x8013186f, "VER_E_LDSTR_OPERAND", nullptr, "corerror.h");
-        ADD(0x80131870, "VER_E_RET_PTR_TO_STACK", nullptr, "corerror.h");
-        ADD(0x80131871, "VER_E_RET_VOID", nullptr, "corerror.h");
-        ADD(0x80131872, "VER_E_RET_MISSING", nullptr, "corerror.h");
-        ADD(0x80131873, "VER_E_RET_EMPTY", nullptr, "corerror.h");
-        ADD(0x80131874, "VER_E_RET_UNINIT", nullptr, "corerror.h");
-        ADD(0x80131875, "VER_E_ARRAY_ACCESS", nullptr, "corerror.h");
-        ADD(0x80131876, "VER_E_ARRAY_V_STORE", nullptr, "corerror.h");
-        ADD(0x80131877, "VER_E_ARRAY_SD", nullptr, "corerror.h");
-        ADD(0x80131878, "VER_E_ARRAY_SD_PTR", nullptr, "corerror.h");
-        ADD(0x80131879, "VER_E_ARRAY_FIELD", nullptr, "corerror.h");
-        ADD(0x8013187a, "VER_E_ARGLIST", nullptr, "corerror.h");
-        ADD(0x8013187b, "VER_E_VALCLASS", nullptr, "corerror.h");
-        ADD(0x8013187c, "VER_E_METHOD_ACCESS", nullptr, "corerror.h");
-        ADD(0x8013187d, "VER_E_FIELD_ACCESS", nullptr, "corerror.h");
-        ADD(0x8013187e, "VER_E_DEAD", nullptr, "corerror.h");
-        ADD(0x8013187f, "VER_E_FIELD_STATIC", nullptr, "corerror.h");
-        ADD(0x80131880, "VER_E_FIELD_NO_STATIC", nullptr, "corerror.h");
-        ADD(0x80131881, "VER_E_ADDR", nullptr, "corerror.h");
-        ADD(0x80131882, "VER_E_ADDR_BYREF", nullptr, "corerror.h");
-        ADD(0x80131883, "VER_E_ADDR_LITERAL", nullptr, "corerror.h");
-        ADD(0x80131884, "VER_E_INITONLY", nullptr, "corerror.h");
-        ADD(0x80131885, "VER_E_THROW", nullptr, "corerror.h");
-        ADD(0x80131886, "VER_E_CALLVIRT_VALCLASS", nullptr, "corerror.h");
-        ADD(0x80131887, "VER_E_CALL_SIG", nullptr, "corerror.h");
-        ADD(0x80131888, "VER_E_CALL_STATIC", nullptr, "corerror.h");
-        ADD(0x80131889, "VER_E_CTOR", nullptr, "corerror.h");
-        ADD(0x8013188a, "VER_E_CTOR_VIRT", nullptr, "corerror.h");
-        ADD(0x8013188b, "VER_E_CTOR_OR_SUPER", nullptr, "corerror.h");
-        ADD(0x8013188c, "VER_E_CTOR_MUL_INIT", nullptr, "corerror.h");
-        ADD(0x8013188d, "VER_E_SIG", nullptr, "corerror.h");
-        ADD(0x8013188e, "VER_E_SIG_ARRAY", nullptr, "corerror.h");
-        ADD(0x8013188f, "VER_E_SIG_ARRAY_PTR", nullptr, "corerror.h");
-        ADD(0x80131890, "VER_E_SIG_ARRAY_BYREF", nullptr, "corerror.h");
-        ADD(0x80131891, "VER_E_SIG_ELEM_PTR", nullptr, "corerror.h");
-        ADD(0x80131892, "VER_E_SIG_VARARG", nullptr, "corerror.h");
-        ADD(0x80131893, "VER_E_SIG_VOID", nullptr, "corerror.h");
-        ADD(0x80131894, "VER_E_SIG_BYREF_BYREF", nullptr, "corerror.h");
-        ADD(0x80131896, "VER_E_CODE_SIZE_ZERO", nullptr, "corerror.h");
-        ADD(0x80131897, "VER_E_BAD_VARARG", nullptr, "corerror.h");
-        ADD(0x80131898, "VER_E_TAIL_CALL", nullptr, "corerror.h");
-        ADD(0x80131899, "VER_E_TAIL_BYREF", nullptr, "corerror.h");
-        ADD(0x8013189a, "VER_E_TAIL_RET", nullptr, "corerror.h");
-        ADD(0x8013189b, "VER_E_TAIL_RET_VOID", nullptr, "corerror.h");
-        ADD(0x8013189c, "VER_E_TAIL_RET_TYPE", nullptr, "corerror.h");
-        ADD(0x8013189d, "VER_E_TAIL_STACK_EMPTY", nullptr, "corerror.h");
-        ADD(0x8013189e, "VER_E_METHOD_END", nullptr, "corerror.h");
-        ADD(0x8013189f, "VER_E_BAD_BRANCH", nullptr, "corerror.h");
-        ADD(0x801318a0, "VER_E_FIN_OVERLAP", nullptr, "corerror.h");
-        ADD(0x801318a1, "VER_E_LEXICAL_NESTING", nullptr, "corerror.h");
-        ADD(0x801318a2, "VER_E_VOLATILE", nullptr, "corerror.h");
-        ADD(0x801318a3, "VER_E_UNALIGNED", nullptr, "corerror.h");
-        ADD(0x801318a4, "VER_E_INNERMOST_FIRST", nullptr, "corerror.h");
-        ADD(0x801318a5, "VER_E_CALLI_VIRTUAL", nullptr, "corerror.h");
-        ADD(0x801318a6, "VER_E_CALL_ABSTRACT", nullptr, "corerror.h");
-        ADD(0x801318a7, "VER_E_STACK_UNEXP_ARRAY", nullptr, "corerror.h");
-        ADD(0x801318a8, "VER_E_NOT_IN_GC_HEAP", nullptr, "corerror.h");
-        ADD(0x801318a9, "VER_E_TRY_N_EMPTY_STACK", nullptr, "corerror.h");
-        ADD(0x801318aa, "VER_E_DLGT_CTOR", nullptr, "corerror.h");
-        ADD(0x801318ab, "VER_E_DLGT_BB", nullptr, "corerror.h");
-        ADD(0x801318ac, "VER_E_DLGT_PATTERN", nullptr, "corerror.h");
-        ADD(0x801318ad, "VER_E_DLGT_LDFTN", nullptr, "corerror.h");
-        ADD(0x801318ae, "VER_E_FTN_ABSTRACT", nullptr, "corerror.h");
-        ADD(0x801318af, "VER_E_SIG_C_VC", nullptr, "corerror.h");
-        ADD(0x801318b0, "VER_E_SIG_VC_C", nullptr, "corerror.h");
-        ADD(0x801318b1, "VER_E_BOX_PTR_TO_STACK", nullptr, "corerror.h");
-        ADD(0x801318b2, "VER_E_SIG_BYREF_TB_AH", nullptr, "corerror.h");
-        ADD(0x801318b3, "VER_E_SIG_ARRAY_TB_AH", nullptr, "corerror.h");
-        ADD(0x801318b4, "VER_E_ENDFILTER_STACK", nullptr, "corerror.h");
-        ADD(0x801318b5, "VER_E_DLGT_SIG_I", nullptr, "corerror.h");
-        ADD(0x801318b6, "VER_E_DLGT_SIG_O", nullptr, "corerror.h");
-        ADD(0x801318b7, "VER_E_RA_PTR_TO_STACK", nullptr, "corerror.h");
-        ADD(0x801318b8, "VER_E_CATCH_VALUE_TYPE", nullptr, "corerror.h");
-        ADD(0x801318b9, "VER_E_CATCH_BYREF", nullptr, "corerror.h");
-        ADD(0x801318ba, "VER_E_FIL_PRECEED_HND", nullptr, "corerror.h");
-        ADD(0x801318bb, "VER_E_LDVIRTFTN_STATIC", nullptr, "corerror.h");
-        ADD(0x801318bc, "VER_E_CALLVIRT_STATIC", nullptr, "corerror.h");
-        ADD(0x801318bd, "VER_E_INITLOCALS", nullptr, "corerror.h");
-        ADD(0x801318be, "VER_E_BR_TO_EXCEPTION", nullptr, "corerror.h");
-        ADD(0x801318bf, "VER_E_CALL_CTOR", nullptr, "corerror.h");
-        ADD(0x801318c0, "VER_E_VALCLASS_OBJREF_VAR", nullptr, "corerror.h");
-        ADD(0x801318c1, "VER_E_STACK_P_VALCLASS_OBJREF_VAR", nullptr, "corerror.h");
-        ADD(0x801318c2, "VER_E_SIG_VAR_PARAM", nullptr, "corerror.h");
-        ADD(0x801318c3, "VER_E_SIG_MVAR_PARAM", nullptr, "corerror.h");
-        ADD(0x801318c4, "VER_E_SIG_VAR_ARG", nullptr, "corerror.h");
-        ADD(0x801318c5, "VER_E_SIG_MVAR_ARG", nullptr, "corerror.h");
-        ADD(0x801318c6, "VER_E_SIG_GENERICINST", nullptr, "corerror.h");
-        ADD(0x801318c7, "VER_E_SIG_METHOD_INST", nullptr, "corerror.h");
-        ADD(0x801318c8, "VER_E_SIG_METHOD_PARENT_INST", nullptr, "corerror.h");
-        ADD(0x801318c9, "VER_E_SIG_FIELD_PARENT_INST", nullptr, "corerror.h");
-        ADD(0x801318ca, "VER_E_CALLCONV_NOT_GENERICINST", nullptr, "corerror.h");
-        ADD(0x801318cb, "VER_E_TOKEN_BAD_METHOD_SPEC", nullptr, "corerror.h");
-        ADD(0x801318cc, "VER_E_BAD_READONLY_PREFIX", nullptr, "corerror.h");
-        ADD(0x801318cd, "VER_E_BAD_CONSTRAINED_PREFIX", nullptr, "corerror.h");
-        ADD(0x801318ce, "VER_E_CIRCULAR_VAR_CONSTRAINTS", nullptr, "corerror.h");
-        ADD(0x801318cf, "VER_E_CIRCULAR_MVAR_CONSTRAINTS", nullptr, "corerror.h");
-        ADD(0x801318d0, "VER_E_UNSATISFIED_METHOD_INST", nullptr, "corerror.h");
-        ADD(0x801318d1, "VER_E_UNSATISFIED_METHOD_PARENT_INST", nullptr, "corerror.h");
-        ADD(0x801318d2, "VER_E_UNSATISFIED_FIELD_PARENT_INST", nullptr, "corerror.h");
-        ADD(0x801318d3, "VER_E_UNSATISFIED_BOX_OPERAND", nullptr, "corerror.h");
-        ADD(0x801318d4, "VER_E_CONSTRAINED_CALL_WITH_NON_BYREF_THIS", nullptr, "corerror.h");
-        ADD(0x801318d5, "VER_E_CONSTRAINED_OF_NON_VARIABLE_TYPE", nullptr, "corerror.h");
-        ADD(0x801318d6, "VER_E_READONLY_UNEXPECTED_CALLEE", nullptr, "corerror.h");
-        ADD(0x801318d7, "VER_E_READONLY_ILLEGAL_WRITE", nullptr, "corerror.h");
-        ADD(0x801318d8, "VER_E_READONLY_IN_MKREFANY", nullptr, "corerror.h");
-        ADD(0x801318d9, "VER_E_UNALIGNED_ALIGNMENT", nullptr, "corerror.h");
-        ADD(0x801318da, "VER_E_TAILCALL_INSIDE_EH", nullptr, "corerror.h");
-        ADD(0x801318db, "VER_E_BACKWARD_BRANCH", nullptr, "corerror.h");
-        ADD(0x801318dc, "VER_E_CALL_TO_VTYPE_BASE", nullptr, "corerror.h");
-        ADD(0x801318dd, "VER_E_NEWOBJ_OF_ABSTRACT_CLASS", nullptr, "corerror.h");
-        ADD(0x801318de, "VER_E_UNMANAGED_POINTER", nullptr, "corerror.h");
-        ADD(0x801318df, "VER_E_LDFTN_NON_FINAL_VIRTUAL", nullptr, "corerror.h");
-        ADD(0x801318e0, "VER_E_FIELD_OVERLAP", nullptr, "corerror.h");
-        ADD(0x801318e1, "VER_E_THIS_MISMATCH", nullptr, "corerror.h");
-        ADD(0x801318e2, "VER_E_STACK_I_I4", nullptr, "corerror.h");
-        ADD(0x801318f0, "VER_E_BAD_PE", nullptr, "corerror.h");
-        ADD(0x801318f1, "VER_E_BAD_MD", nullptr, "corerror.h");
-        ADD(0x801318f2, "VER_E_BAD_APPDOMAIN", nullptr, "corerror.h");
-        ADD(0x801318f3, "VER_E_TYPELOAD", nullptr, "corerror.h");
-        ADD(0x801318f4, "VER_E_PE_LOAD", nullptr, "corerror.h");
-        ADD(0x801318f5, "VER_E_WRITE_RVA_STATIC", nullptr, "corerror.h");
+        ADD(0x80131801, "VER_E_HRESULT", nil, "corerror.h");
+        ADD(0x80131802, "VER_E_OFFSET", nil, "corerror.h");
+        ADD(0x80131803, "VER_E_OPCODE", nil, "corerror.h");
+        ADD(0x80131804, "VER_E_OPERAND", nil, "corerror.h");
+        ADD(0x80131805, "VER_E_TOKEN", nil, "corerror.h");
+        ADD(0x80131806, "VER_E_EXCEPT", nil, "corerror.h");
+        ADD(0x80131807, "VER_E_STACK_SLOT", nil, "corerror.h");
+        ADD(0x80131808, "VER_E_LOC", nil, "corerror.h");
+        ADD(0x80131809, "VER_E_ARG", nil, "corerror.h");
+        ADD(0x8013180a, "VER_E_FOUND", nil, "corerror.h");
+        ADD(0x8013180b, "VER_E_EXPECTED", nil, "corerror.h");
+        ADD(0x8013180c, "VER_E_LOC_BYNAME", nil, "corerror.h");
+        ADD(0x80131810, "VER_E_UNKNOWN_OPCODE", nil, "corerror.h");
+        ADD(0x80131811, "VER_E_SIG_CALLCONV", nil, "corerror.h");
+        ADD(0x80131812, "VER_E_SIG_ELEMTYPE", nil, "corerror.h");
+        ADD(0x80131814, "VER_E_RET_SIG", nil, "corerror.h");
+        ADD(0x80131815, "VER_E_FIELD_SIG", nil, "corerror.h");
+        ADD(0x80131818, "VER_E_INTERNAL", nil, "corerror.h");
+        ADD(0x80131819, "VER_E_STACK_TOO_LARGE", nil, "corerror.h");
+        ADD(0x8013181a, "VER_E_ARRAY_NAME_LONG", nil, "corerror.h");
+        ADD(0x80131820, "VER_E_FALLTHRU", nil, "corerror.h");
+        ADD(0x80131821, "VER_E_TRY_GTEQ_END", nil, "corerror.h");
+        ADD(0x80131822, "VER_E_TRYEND_GT_CS", nil, "corerror.h");
+        ADD(0x80131823, "VER_E_HND_GTEQ_END", nil, "corerror.h");
+        ADD(0x80131824, "VER_E_HNDEND_GT_CS", nil, "corerror.h");
+        ADD(0x80131825, "VER_E_FLT_GTEQ_CS", nil, "corerror.h");
+        ADD(0x80131826, "VER_E_TRY_START", nil, "corerror.h");
+        ADD(0x80131827, "VER_E_HND_START", nil, "corerror.h");
+        ADD(0x80131828, "VER_E_FLT_START", nil, "corerror.h");
+        ADD(0x80131829, "VER_E_TRY_OVERLAP", nil, "corerror.h");
+        ADD(0x8013182a, "VER_E_TRY_EQ_HND_FIL", nil, "corerror.h");
+        ADD(0x8013182b, "VER_E_TRY_SHARE_FIN_FAL", nil, "corerror.h");
+        ADD(0x8013182c, "VER_E_HND_OVERLAP", nil, "corerror.h");
+        ADD(0x8013182d, "VER_E_HND_EQ", nil, "corerror.h");
+        ADD(0x8013182e, "VER_E_FIL_OVERLAP", nil, "corerror.h");
+        ADD(0x8013182f, "VER_E_FIL_EQ", nil, "corerror.h");
+        ADD(0x80131830, "VER_E_FIL_CONT_TRY", nil, "corerror.h");
+        ADD(0x80131831, "VER_E_FIL_CONT_HND", nil, "corerror.h");
+        ADD(0x80131832, "VER_E_FIL_CONT_FIL", nil, "corerror.h");
+        ADD(0x80131833, "VER_E_FIL_GTEQ_CS", nil, "corerror.h");
+        ADD(0x80131834, "VER_E_FIL_START", nil, "corerror.h");
+        ADD(0x80131835, "VER_E_FALLTHRU_EXCEP", nil, "corerror.h");
+        ADD(0x80131836, "VER_E_FALLTHRU_INTO_HND", nil, "corerror.h");
+        ADD(0x80131837, "VER_E_FALLTHRU_INTO_FIL", nil, "corerror.h");
+        ADD(0x80131838, "VER_E_LEAVE", nil, "corerror.h");
+        ADD(0x80131839, "VER_E_RETHROW", nil, "corerror.h");
+        ADD(0x8013183a, "VER_E_ENDFINALLY", nil, "corerror.h");
+        ADD(0x8013183b, "VER_E_ENDFILTER", nil, "corerror.h");
+        ADD(0x8013183c, "VER_E_ENDFILTER_MISSING", nil, "corerror.h");
+        ADD(0x8013183d, "VER_E_BR_INTO_TRY", nil, "corerror.h");
+        ADD(0x8013183e, "VER_E_BR_INTO_HND", nil, "corerror.h");
+        ADD(0x8013183f, "VER_E_BR_INTO_FIL", nil, "corerror.h");
+        ADD(0x80131840, "VER_E_BR_OUTOF_TRY", nil, "corerror.h");
+        ADD(0x80131841, "VER_E_BR_OUTOF_HND", nil, "corerror.h");
+        ADD(0x80131842, "VER_E_BR_OUTOF_FIL", nil, "corerror.h");
+        ADD(0x80131843, "VER_E_BR_OUTOF_FIN", nil, "corerror.h");
+        ADD(0x80131844, "VER_E_RET_FROM_TRY", nil, "corerror.h");
+        ADD(0x80131845, "VER_E_RET_FROM_HND", nil, "corerror.h");
+        ADD(0x80131846, "VER_E_RET_FROM_FIL", nil, "corerror.h");
+        ADD(0x80131847, "VER_E_BAD_JMP_TARGET", nil, "corerror.h");
+        ADD(0x80131848, "VER_E_PATH_LOC", nil, "corerror.h");
+        ADD(0x80131849, "VER_E_PATH_THIS", nil, "corerror.h");
+        ADD(0x8013184a, "VER_E_PATH_STACK", nil, "corerror.h");
+        ADD(0x8013184b, "VER_E_PATH_STACK_DEPTH", nil, "corerror.h");
+        ADD(0x8013184c, "VER_E_THIS", nil, "corerror.h");
+        ADD(0x8013184d, "VER_E_THIS_UNINIT_EXCEP", nil, "corerror.h");
+        ADD(0x8013184e, "VER_E_THIS_UNINIT_STORE", nil, "corerror.h");
+        ADD(0x8013184f, "VER_E_THIS_UNINIT_RET", nil, "corerror.h");
+        ADD(0x80131850, "VER_E_THIS_UNINIT_V_RET", nil, "corerror.h");
+        ADD(0x80131851, "VER_E_THIS_UNINIT_BR", nil, "corerror.h");
+        ADD(0x80131852, "VER_E_LDFTN_CTOR", nil, "corerror.h");
+        ADD(0x80131853, "VER_E_STACK_NOT_EQ", nil, "corerror.h");
+        ADD(0x80131854, "VER_E_STACK_UNEXPECTED", nil, "corerror.h");
+        ADD(0x80131855, "VER_E_STACK_EXCEPTION", nil, "corerror.h");
+        ADD(0x80131856, "VER_E_STACK_OVERFLOW", nil, "corerror.h");
+        ADD(0x80131857, "VER_E_STACK_UNDERFLOW", nil, "corerror.h");
+        ADD(0x80131858, "VER_E_STACK_EMPTY", nil, "corerror.h");
+        ADD(0x80131859, "VER_E_STACK_UNINIT", nil, "corerror.h");
+        ADD(0x8013185a, "VER_E_STACK_I_I4_I8", nil, "corerror.h");
+        ADD(0x8013185b, "VER_E_STACK_R_R4_R8", nil, "corerror.h");
+        ADD(0x8013185c, "VER_E_STACK_NO_R_I8", nil, "corerror.h");
+        ADD(0x8013185d, "VER_E_STACK_NUMERIC", nil, "corerror.h");
+        ADD(0x8013185e, "VER_E_STACK_OBJREF", nil, "corerror.h");
+        ADD(0x8013185f, "VER_E_STACK_P_OBJREF", nil, "corerror.h");
+        ADD(0x80131860, "VER_E_STACK_BYREF", nil, "corerror.h");
+        ADD(0x80131861, "VER_E_STACK_METHOD", nil, "corerror.h");
+        ADD(0x80131862, "VER_E_STACK_ARRAY_SD", nil, "corerror.h");
+        ADD(0x80131863, "VER_E_STACK_VALCLASS", nil, "corerror.h");
+        ADD(0x80131864, "VER_E_STACK_P_VALCLASS", nil, "corerror.h");
+        ADD(0x80131865, "VER_E_STACK_NO_VALCLASS", nil, "corerror.h");
+        ADD(0x80131866, "VER_E_LOC_DEAD", nil, "corerror.h");
+        ADD(0x80131867, "VER_E_LOC_NUM", nil, "corerror.h");
+        ADD(0x80131868, "VER_E_ARG_NUM", nil, "corerror.h");
+        ADD(0x80131869, "VER_E_TOKEN_RESOLVE", nil, "corerror.h");
+        ADD(0x8013186a, "VER_E_TOKEN_TYPE", nil, "corerror.h");
+        ADD(0x8013186b, "VER_E_TOKEN_TYPE_MEMBER", nil, "corerror.h");
+        ADD(0x8013186c, "VER_E_TOKEN_TYPE_FIELD", nil, "corerror.h");
+        ADD(0x8013186d, "VER_E_TOKEN_TYPE_SIG", nil, "corerror.h");
+        ADD(0x8013186e, "VER_E_UNVERIFIABLE", nil, "corerror.h");
+        ADD(0x8013186f, "VER_E_LDSTR_OPERAND", nil, "corerror.h");
+        ADD(0x80131870, "VER_E_RET_PTR_TO_STACK", nil, "corerror.h");
+        ADD(0x80131871, "VER_E_RET_VOID", nil, "corerror.h");
+        ADD(0x80131872, "VER_E_RET_MISSING", nil, "corerror.h");
+        ADD(0x80131873, "VER_E_RET_EMPTY", nil, "corerror.h");
+        ADD(0x80131874, "VER_E_RET_UNINIT", nil, "corerror.h");
+        ADD(0x80131875, "VER_E_ARRAY_ACCESS", nil, "corerror.h");
+        ADD(0x80131876, "VER_E_ARRAY_V_STORE", nil, "corerror.h");
+        ADD(0x80131877, "VER_E_ARRAY_SD", nil, "corerror.h");
+        ADD(0x80131878, "VER_E_ARRAY_SD_PTR", nil, "corerror.h");
+        ADD(0x80131879, "VER_E_ARRAY_FIELD", nil, "corerror.h");
+        ADD(0x8013187a, "VER_E_ARGLIST", nil, "corerror.h");
+        ADD(0x8013187b, "VER_E_VALCLASS", nil, "corerror.h");
+        ADD(0x8013187c, "VER_E_METHOD_ACCESS", nil, "corerror.h");
+        ADD(0x8013187d, "VER_E_FIELD_ACCESS", nil, "corerror.h");
+        ADD(0x8013187e, "VER_E_DEAD", nil, "corerror.h");
+        ADD(0x8013187f, "VER_E_FIELD_STATIC", nil, "corerror.h");
+        ADD(0x80131880, "VER_E_FIELD_NO_STATIC", nil, "corerror.h");
+        ADD(0x80131881, "VER_E_ADDR", nil, "corerror.h");
+        ADD(0x80131882, "VER_E_ADDR_BYREF", nil, "corerror.h");
+        ADD(0x80131883, "VER_E_ADDR_LITERAL", nil, "corerror.h");
+        ADD(0x80131884, "VER_E_INITONLY", nil, "corerror.h");
+        ADD(0x80131885, "VER_E_THROW", nil, "corerror.h");
+        ADD(0x80131886, "VER_E_CALLVIRT_VALCLASS", nil, "corerror.h");
+        ADD(0x80131887, "VER_E_CALL_SIG", nil, "corerror.h");
+        ADD(0x80131888, "VER_E_CALL_STATIC", nil, "corerror.h");
+        ADD(0x80131889, "VER_E_CTOR", nil, "corerror.h");
+        ADD(0x8013188a, "VER_E_CTOR_VIRT", nil, "corerror.h");
+        ADD(0x8013188b, "VER_E_CTOR_OR_SUPER", nil, "corerror.h");
+        ADD(0x8013188c, "VER_E_CTOR_MUL_INIT", nil, "corerror.h");
+        ADD(0x8013188d, "VER_E_SIG", nil, "corerror.h");
+        ADD(0x8013188e, "VER_E_SIG_ARRAY", nil, "corerror.h");
+        ADD(0x8013188f, "VER_E_SIG_ARRAY_PTR", nil, "corerror.h");
+        ADD(0x80131890, "VER_E_SIG_ARRAY_BYREF", nil, "corerror.h");
+        ADD(0x80131891, "VER_E_SIG_ELEM_PTR", nil, "corerror.h");
+        ADD(0x80131892, "VER_E_SIG_VARARG", nil, "corerror.h");
+        ADD(0x80131893, "VER_E_SIG_VOID", nil, "corerror.h");
+        ADD(0x80131894, "VER_E_SIG_BYREF_BYREF", nil, "corerror.h");
+        ADD(0x80131896, "VER_E_CODE_SIZE_ZERO", nil, "corerror.h");
+        ADD(0x80131897, "VER_E_BAD_VARARG", nil, "corerror.h");
+        ADD(0x80131898, "VER_E_TAIL_CALL", nil, "corerror.h");
+        ADD(0x80131899, "VER_E_TAIL_BYREF", nil, "corerror.h");
+        ADD(0x8013189a, "VER_E_TAIL_RET", nil, "corerror.h");
+        ADD(0x8013189b, "VER_E_TAIL_RET_VOID", nil, "corerror.h");
+        ADD(0x8013189c, "VER_E_TAIL_RET_TYPE", nil, "corerror.h");
+        ADD(0x8013189d, "VER_E_TAIL_STACK_EMPTY", nil, "corerror.h");
+        ADD(0x8013189e, "VER_E_METHOD_END", nil, "corerror.h");
+        ADD(0x8013189f, "VER_E_BAD_BRANCH", nil, "corerror.h");
+        ADD(0x801318a0, "VER_E_FIN_OVERLAP", nil, "corerror.h");
+        ADD(0x801318a1, "VER_E_LEXICAL_NESTING", nil, "corerror.h");
+        ADD(0x801318a2, "VER_E_VOLATILE", nil, "corerror.h");
+        ADD(0x801318a3, "VER_E_UNALIGNED", nil, "corerror.h");
+        ADD(0x801318a4, "VER_E_INNERMOST_FIRST", nil, "corerror.h");
+        ADD(0x801318a5, "VER_E_CALLI_VIRTUAL", nil, "corerror.h");
+        ADD(0x801318a6, "VER_E_CALL_ABSTRACT", nil, "corerror.h");
+        ADD(0x801318a7, "VER_E_STACK_UNEXP_ARRAY", nil, "corerror.h");
+        ADD(0x801318a8, "VER_E_NOT_IN_GC_HEAP", nil, "corerror.h");
+        ADD(0x801318a9, "VER_E_TRY_N_EMPTY_STACK", nil, "corerror.h");
+        ADD(0x801318aa, "VER_E_DLGT_CTOR", nil, "corerror.h");
+        ADD(0x801318ab, "VER_E_DLGT_BB", nil, "corerror.h");
+        ADD(0x801318ac, "VER_E_DLGT_PATTERN", nil, "corerror.h");
+        ADD(0x801318ad, "VER_E_DLGT_LDFTN", nil, "corerror.h");
+        ADD(0x801318ae, "VER_E_FTN_ABSTRACT", nil, "corerror.h");
+        ADD(0x801318af, "VER_E_SIG_C_VC", nil, "corerror.h");
+        ADD(0x801318b0, "VER_E_SIG_VC_C", nil, "corerror.h");
+        ADD(0x801318b1, "VER_E_BOX_PTR_TO_STACK", nil, "corerror.h");
+        ADD(0x801318b2, "VER_E_SIG_BYREF_TB_AH", nil, "corerror.h");
+        ADD(0x801318b3, "VER_E_SIG_ARRAY_TB_AH", nil, "corerror.h");
+        ADD(0x801318b4, "VER_E_ENDFILTER_STACK", nil, "corerror.h");
+        ADD(0x801318b5, "VER_E_DLGT_SIG_I", nil, "corerror.h");
+        ADD(0x801318b6, "VER_E_DLGT_SIG_O", nil, "corerror.h");
+        ADD(0x801318b7, "VER_E_RA_PTR_TO_STACK", nil, "corerror.h");
+        ADD(0x801318b8, "VER_E_CATCH_VALUE_TYPE", nil, "corerror.h");
+        ADD(0x801318b9, "VER_E_CATCH_BYREF", nil, "corerror.h");
+        ADD(0x801318ba, "VER_E_FIL_PRECEED_HND", nil, "corerror.h");
+        ADD(0x801318bb, "VER_E_LDVIRTFTN_STATIC", nil, "corerror.h");
+        ADD(0x801318bc, "VER_E_CALLVIRT_STATIC", nil, "corerror.h");
+        ADD(0x801318bd, "VER_E_INITLOCALS", nil, "corerror.h");
+        ADD(0x801318be, "VER_E_BR_TO_EXCEPTION", nil, "corerror.h");
+        ADD(0x801318bf, "VER_E_CALL_CTOR", nil, "corerror.h");
+        ADD(0x801318c0, "VER_E_VALCLASS_OBJREF_VAR", nil, "corerror.h");
+        ADD(0x801318c1, "VER_E_STACK_P_VALCLASS_OBJREF_VAR", nil, "corerror.h");
+        ADD(0x801318c2, "VER_E_SIG_VAR_PARAM", nil, "corerror.h");
+        ADD(0x801318c3, "VER_E_SIG_MVAR_PARAM", nil, "corerror.h");
+        ADD(0x801318c4, "VER_E_SIG_VAR_ARG", nil, "corerror.h");
+        ADD(0x801318c5, "VER_E_SIG_MVAR_ARG", nil, "corerror.h");
+        ADD(0x801318c6, "VER_E_SIG_GENERICINST", nil, "corerror.h");
+        ADD(0x801318c7, "VER_E_SIG_METHOD_INST", nil, "corerror.h");
+        ADD(0x801318c8, "VER_E_SIG_METHOD_PARENT_INST", nil, "corerror.h");
+        ADD(0x801318c9, "VER_E_SIG_FIELD_PARENT_INST", nil, "corerror.h");
+        ADD(0x801318ca, "VER_E_CALLCONV_NOT_GENERICINST", nil, "corerror.h");
+        ADD(0x801318cb, "VER_E_TOKEN_BAD_METHOD_SPEC", nil, "corerror.h");
+        ADD(0x801318cc, "VER_E_BAD_READONLY_PREFIX", nil, "corerror.h");
+        ADD(0x801318cd, "VER_E_BAD_CONSTRAINED_PREFIX", nil, "corerror.h");
+        ADD(0x801318ce, "VER_E_CIRCULAR_VAR_CONSTRAINTS", nil, "corerror.h");
+        ADD(0x801318cf, "VER_E_CIRCULAR_MVAR_CONSTRAINTS", nil, "corerror.h");
+        ADD(0x801318d0, "VER_E_UNSATISFIED_METHOD_INST", nil, "corerror.h");
+        ADD(0x801318d1, "VER_E_UNSATISFIED_METHOD_PARENT_INST", nil, "corerror.h");
+        ADD(0x801318d2, "VER_E_UNSATISFIED_FIELD_PARENT_INST", nil, "corerror.h");
+        ADD(0x801318d3, "VER_E_UNSATISFIED_BOX_OPERAND", nil, "corerror.h");
+        ADD(0x801318d4, "VER_E_CONSTRAINED_CALL_WITH_NON_BYREF_THIS", nil, "corerror.h");
+        ADD(0x801318d5, "VER_E_CONSTRAINED_OF_NON_VARIABLE_TYPE", nil, "corerror.h");
+        ADD(0x801318d6, "VER_E_READONLY_UNEXPECTED_CALLEE", nil, "corerror.h");
+        ADD(0x801318d7, "VER_E_READONLY_ILLEGAL_WRITE", nil, "corerror.h");
+        ADD(0x801318d8, "VER_E_READONLY_IN_MKREFANY", nil, "corerror.h");
+        ADD(0x801318d9, "VER_E_UNALIGNED_ALIGNMENT", nil, "corerror.h");
+        ADD(0x801318da, "VER_E_TAILCALL_INSIDE_EH", nil, "corerror.h");
+        ADD(0x801318db, "VER_E_BACKWARD_BRANCH", nil, "corerror.h");
+        ADD(0x801318dc, "VER_E_CALL_TO_VTYPE_BASE", nil, "corerror.h");
+        ADD(0x801318dd, "VER_E_NEWOBJ_OF_ABSTRACT_CLASS", nil, "corerror.h");
+        ADD(0x801318de, "VER_E_UNMANAGED_POINTER", nil, "corerror.h");
+        ADD(0x801318df, "VER_E_LDFTN_NON_FINAL_VIRTUAL", nil, "corerror.h");
+        ADD(0x801318e0, "VER_E_FIELD_OVERLAP", nil, "corerror.h");
+        ADD(0x801318e1, "VER_E_THIS_MISMATCH", nil, "corerror.h");
+        ADD(0x801318e2, "VER_E_STACK_I_I4", nil, "corerror.h");
+        ADD(0x801318f0, "VER_E_BAD_PE", nil, "corerror.h");
+        ADD(0x801318f1, "VER_E_BAD_MD", nil, "corerror.h");
+        ADD(0x801318f2, "VER_E_BAD_APPDOMAIN", nil, "corerror.h");
+        ADD(0x801318f3, "VER_E_TYPELOAD", nil, "corerror.h");
+        ADD(0x801318f4, "VER_E_PE_LOAD", nil, "corerror.h");
+        ADD(0x801318f5, "VER_E_WRITE_RVA_STATIC", nil, "corerror.h");
         ADD(0x80131904, "COR_E_SqlException", "System.Data.SqlClient.SqlClientException", "corerror.h");
-        ADD(0x80131920, "COR_E_Data", nullptr, "corerror.h");
-        ADD(0x80131921, "COR_E_DataDeletedRowInaccessible", nullptr, "corerror.h");
-        ADD(0x80131922, "COR_E_DataDuplicateName", nullptr, "corerror.h");
-        ADD(0x80131923, "COR_E_DataInRowChangingEvent", nullptr, "corerror.h");
-        ADD(0x80131924, "COR_E_DataInvalidConstraint", nullptr, "corerror.h");
-        ADD(0x80131925, "COR_E_DataMissingPrimaryKey", nullptr, "corerror.h");
-        ADD(0x80131926, "COR_E_DataNoNullAllowed", nullptr, "corerror.h");
-        ADD(0x80131927, "COR_E_DataReadOnly", nullptr, "corerror.h");
-        ADD(0x80131928, "COR_E_DataRowNotInTable", nullptr, "corerror.h");
-        ADD(0x80131929, "COR_E_DataVersionNotFound", nullptr, "corerror.h");
-        ADD(0x8013192a, "COR_E_DataConstraint", nullptr, "corerror.h");
-        ADD(0x8013192b, "COR_E_StrongTyping", nullptr, "corerror.h");
-        ADD(0x80131930, "COR_E_SqlType", nullptr, "corerror.h");
-        ADD(0x80131931, "COR_E_SqlNullValue", nullptr, "corerror.h");
-        ADD(0x80131932, "COR_E_SqlTruncate", nullptr, "corerror.h");
-        ADD(0x80131933, "COR_E_AdapterMapping", nullptr, "corerror.h");
-        ADD(0x80131934, "COR_E_DataAdapter", nullptr, "corerror.h");
-        ADD(0x80131935, "COR_E_DBConcurrency", nullptr, "corerror.h");
-        ADD(0x80131936, "COR_E_OperationAborted", nullptr, "corerror.h");
-        ADD(0x80131937, "COR_E_InvalidUdt", nullptr, "corerror.h");
+        ADD(0x80131920, "COR_E_Data", nil, "corerror.h");
+        ADD(0x80131921, "COR_E_DataDeletedRowInaccessible", nil, "corerror.h");
+        ADD(0x80131922, "COR_E_DataDuplicateName", nil, "corerror.h");
+        ADD(0x80131923, "COR_E_DataInRowChangingEvent", nil, "corerror.h");
+        ADD(0x80131924, "COR_E_DataInvalidConstraint", nil, "corerror.h");
+        ADD(0x80131925, "COR_E_DataMissingPrimaryKey", nil, "corerror.h");
+        ADD(0x80131926, "COR_E_DataNoNullAllowed", nil, "corerror.h");
+        ADD(0x80131927, "COR_E_DataReadOnly", nil, "corerror.h");
+        ADD(0x80131928, "COR_E_DataRowNotInTable", nil, "corerror.h");
+        ADD(0x80131929, "COR_E_DataVersionNotFound", nil, "corerror.h");
+        ADD(0x8013192a, "COR_E_DataConstraint", nil, "corerror.h");
+        ADD(0x8013192b, "COR_E_StrongTyping", nil, "corerror.h");
+        ADD(0x80131930, "COR_E_SqlType", nil, "corerror.h");
+        ADD(0x80131931, "COR_E_SqlNullValue", nil, "corerror.h");
+        ADD(0x80131932, "COR_E_SqlTruncate", nil, "corerror.h");
+        ADD(0x80131933, "COR_E_AdapterMapping", nil, "corerror.h");
+        ADD(0x80131934, "COR_E_DataAdapter", nil, "corerror.h");
+        ADD(0x80131935, "COR_E_DBConcurrency", nil, "corerror.h");
+        ADD(0x80131936, "COR_E_OperationAborted", nil, "corerror.h");
+        ADD(0x80131937, "COR_E_InvalidUdt", nil, "corerror.h");
         ADD(0x80131937, "COR_E_OdbcException", "System.Data.Odbc.OdbcException", "corerror.h");
         ADD(0x80131938, "COR_E_OracleException", "System.Data.OracleClient.OracleException", "corerror.h");
-        ADD(0x80131940, "COR_E_Xml", nullptr, "corerror.h");
-        ADD(0x80131941, "COR_E_XmlSchema", nullptr, "corerror.h");
-        ADD(0x80131942, "COR_E_XmlXslt", nullptr, "corerror.h");
-        ADD(0x80131943, "COR_E_XmlXPath", nullptr, "corerror.h");
-        ADD(0x80131944, "COR_E_XmlQuery", nullptr, "corerror.h");
+        ADD(0x80131940, "COR_E_Xml", nil, "corerror.h");
+        ADD(0x80131941, "COR_E_XmlSchema", nil, "corerror.h");
+        ADD(0x80131942, "COR_E_XmlXslt", nil, "corerror.h");
+        ADD(0x80131943, "COR_E_XmlXPath", nil, "corerror.h");
+        ADD(0x80131944, "COR_E_XmlQuery", nil, "corerror.h");
         ADD(0x80131b00, "VLDTR_E_IFACE_NOTIFACE", "Interface in InterfaceImpl is not marked tdInterface", "corerror.h");
         ADD(0x80131b01, "VLDTR_E_FD_RVAHASNORVA", "Field marked fdHasFieldRVA but has no RVA rec", "corerror.h");
         ADD(0x80131b02, "VLDTR_E_FD_RVAHASZERORVA", "Field marked fdHasFieldRVA has RVA =0", "corerror.h");
@@ -1023,8 +1023,8 @@ namespace common::errormsg::cor {
         ADD(0x80131b07, "VLDTR_E_TD_EXTTYPESPEC", "Class extends TypeSpec (warning)", "corerror.h");
         ADD(0x80131b09, "VLDTR_E_TD_VTNOSIZE", "Value Class has zero size", "corerror.h");
         ADD(0x80131b0a, "VLDTR_E_TD_IFACESEALED", "Interface is sealed", "corerror.h");
-        ADD(0x80131b0b, "VLDTR_E_NC_BADNESTED", "Bad nullptrnestednullptr token in NestedClass", "corerror.h");
-        ADD(0x80131b0c, "VLDTR_E_NC_BADENCLOSER", "Bad nullptrenclosingnullptr token in NestedClass", "corerror.h");
+        ADD(0x80131b0b, "VLDTR_E_NC_BADNESTED", "Bad \"nested\" token in NestedClass", "corerror.h");
+        ADD(0x80131b0c, "VLDTR_E_NC_BADENCLOSER", "Bad \"enclosing\" token in NestedClass", "corerror.h");
         ADD(0x80131b0d, "VLDTR_E_NC_DUP", "Duplicate NestedClass record", "corerror.h");
         ADD(0x80131b0e, "VLDTR_E_NC_DUPENCLOSER", "Duplicate NestedClass with different encloser", "corerror.h");
         ADD(0x80131b0f, "VLDTR_E_FRVA_ZERORVA", "RVA=0 in FieldRVA record", "corerror.h");
@@ -1048,7 +1048,7 @@ namespace common::errormsg::cor {
         ADD(0x80131b21, "VLDTR_E_EP_TOOMANYARGS", "Entry point has more than one arg", "corerror.h");
         ADD(0x80131b22, "VLDTR_E_EP_BADRET", "Entry point has bad return type", "corerror.h");
         ADD(0x80131b23, "VLDTR_E_EP_BADARG", "Entry point has bad argument", "corerror.h");
-        ADD(0x80131b24, "VLDTR_E_SIG_BADVOID", "Illegal nullptrvoidnullptr in signature", "corerror.h");
+        ADD(0x80131b24, "VLDTR_E_SIG_BADVOID", "Illegal \"void\" in signature", "corerror.h");
         ADD(0x80131b25, "VLDTR_E_IFACE_METHMULTIMPL", "Multiple implementation of method", "corerror.h");
         ADD(0x80131b26, "VLDTR_E_GP_NAMENULL", "GenericParam name is NULL", "corerror.h");
         ADD(0x80131b27, "VLDTR_E_GP_OWNERNIL", "GenericParam has nil owner.", "corerror.h");
@@ -1128,7 +1128,7 @@ namespace common::errormsg::cor {
         ADD(0x80131c24, "CORDBG_E_ILLEGAL_IN_PROLOG", "The operation failed because the thread is in the prolog", "corerror.h");
         ADD(0x80131c25, "CORDBG_E_ILLEGAL_IN_NATIVE_CODE", "The operation failed because the thread is in native code", "corerror.h");
         ADD(0x80131c26, "CORDBG_E_ILLEGAL_IN_OPTIMIZED_CODE", "The operation failed because the thread is in optimized code.", "corerror.h");
-        ADD(0x80131c27, "CORDBG_E_MINIDUMP_UNSUPPORTED", nullptr, "corerror.h");
+        ADD(0x80131c27, "CORDBG_E_MINIDUMP_UNSUPPORTED", nil, "corerror.h");
         ADD(0x80131c28, "CORDBG_E_APPDOMAIN_MISMATCH", "A supplied object or type belongs to the wrong AppDomain", "corerror.h");
         ADD(0x80131c29, "CORDBG_E_CONTEXT_UNVAILABLE", "The thread's context is not available.", "corerror.h");
         ADD(0x80131c30, "CORDBG_E_UNCOMPATIBLE_PLATFORMS", "The operation failed because debuggee and debugger are on incompatible platform", "corerror.h");
@@ -1150,7 +1150,7 @@ namespace common::errormsg::cor {
         ADD(0x80131e00, "CLR_OPTSVC_E_CONTROLLER_INTERRUPT", "Service controller interrupted work", "corerror.h");
         ADD(0x80131f00, "NGEN_FAILED_GET_DEPENDENCIES", "Service manager failed to get ICorSvcDependencies interface", "corerror.h");
         ADD(0x80131f01, "NGEN_FAILED_NATIVE_IMAGE_DELETE", "Failed to delete native image", "corerror.h");
-        ADD(0x80131fff, "CLDB_E_INTERNALERROR", nullptr, "corerror.h");
+        ADD(0x80131fff, "CLDB_E_INTERNALERROR", nil, "corerror.h");
     };
     ON_INIT{
         for (size_t i = 1; i < messages.size(); i++) {
