@@ -7,7 +7,7 @@ using namespace std;
 
 namespace common::helper::encoding {
     wstring ConvertToUtf16(const char* utf8str) {
-        auto chrCount = MultiByteToWideChar(CP_UTF8, 0, utf8str, -1, nullptr, 0);
+        auto chrCount = MultiByteToWideChar(CP_UTF8, 0, utf8str, -1, nil, 0);
         if (chrCount == 0)
             return wstring();
         wstring output(chrCount - 1, '\0');
@@ -16,11 +16,11 @@ namespace common::helper::encoding {
     }
 
     string ConvertToUtf8(const wchar_t* utf16str) {
-        auto byteCount = WideCharToMultiByte(CP_UTF8, 0, utf16str, -1, nullptr, 0, nullptr, nullptr);
+        auto byteCount = WideCharToMultiByte(CP_UTF8, 0, utf16str, -1, nil, 0, nil, nil);
         if (byteCount == 0)
             return string();
         string output(byteCount - 1, '\0');
-        WideCharToMultiByte(CP_UTF8, 0, utf16str, -1, output.data(), byteCount, nullptr, nullptr);
+        WideCharToMultiByte(CP_UTF8, 0, utf16str, -1, output.data(), byteCount, nil, nil);
         return output;
     }
 }
