@@ -1,10 +1,10 @@
 #pragma once
 #include "framework.h"
 #include <string>
+#include <span>
+#include <functional>
 
 namespace common::helper::memory {
-    using ImportTableCallbackType = void (*)(LPCSTR importDllName);
-    DWORD ResolveAddress(DWORD* offsets, int length);
+    DWORD ResolveAddress(std::span<const DWORD> offsets);
     std::string GetAddressConfigAsString();
-    void ScanImportTable(HMODULE hModule, ImportTableCallbackType callback);
 }
