@@ -19,12 +19,18 @@ namespace errormsg = common::errormsg;
 using namespace std;
 using namespace Microsoft::WRL;
 
-tm& GetTimeNow() {
-    auto now = time(nullptr);
-    return *localtime(&now);
-}
-
 namespace common::log {
+    FILE* __logFile;
+    wstring __logPath;
+    string __processName;
+
+    tm GetTimeNow() {
+        tm rs;
+        auto now = time(nil);
+        localtime_s(&rs, &now);
+        return rs;
+    }
+
     void OpenConsole() {
         if (AllocConsole() == FALSE)
             return;
@@ -37,41 +43,14 @@ namespace common::log {
         printf("Debugging Window:\n\n");
     }
 
-    FILE* logFile;
-    wstring logPath;
-    string processName;
-    void ToFile(const char* _Format, ...) {
-        va_list args;
-        va_start(args, _Format);
-        if (!logFile) {
-            if (logPath.size() == 0) {
-                logPath = wstring(g_currentModuleDirPath) + L"/log.txt";
-                processName = encoding::ConvertToUtf8(g_currentConfig.ProcessName);
-            }
-            logFile = _wfsopen(logPath.c_str(), L"a+", _SH_DENYNO);
-            if (logFile != NULL)
-                setvbuf(logFile, NULL, _IONBF, 0);
-        }
-        if (logFile != NULL) {
-            auto& now = GetTimeNow();
-            fprintf(logFile, "[%s %02d/%02d/%02d %02d:%02d:%02d] ",
-                processName.c_str(),
-                now.tm_mday, now.tm_mon + 1, now.tm_year + 1900,
-                now.tm_hour, now.tm_min, now.tm_sec);
-            vfprintf(logFile, _Format, args);
-            fprintf(logFile, "\n");
-        }
-        va_end(args);
-    }
-
     void DxErrToFile(const char* message, HRESULT hResult) {
         auto errorStr = DXGetErrorStringA(hResult);
-        if (errorStr == NULL) {
+        if (errorStr == nil) {
             HResultToFile(message, hResult);
             return;
         }
         auto errorDes = DXGetErrorDescriptionA(hResult);
-        auto description = errorDes != NULL ? string(errorStr) + ": " + errorDes : string(errorStr);
+        auto description = errorDes != nil ? string(errorStr) + ": " + errorDes : string(errorStr);
 #if _DEBUG
         ToConsole("%s: %s", message, description.c_str());
 #endif
@@ -80,14 +59,14 @@ namespace common::log {
 
     void HResultToFile(const char* message, HRESULT hResult) {
         ComPtr<IErrorInfo> errorInfo;
-        auto _ = GetErrorInfo(0, &errorInfo);
+        std::ignore = GetErrorInfo(0, &errorInfo);
         _com_error error(hResult, errorInfo.Get(), true);
         auto description = error.Description();
         if (description.length() > 0) {
 #if _DEBUG
-            ToConsole("%s: %s", message, description);
+            ToConsole("%s: %s", message, (char*)description);
 #endif
-            ToFile("%s: %s", message, description);
+            ToFile("%s: %s", message, (char*)description);
             return;
         }
         auto errorMessage = string(error.ErrorMessage());
@@ -99,7 +78,7 @@ namespace common::log {
             errorMessage = errormsg::GuessErrorsFromHResult(hResult);
             if (errorMessage != "") {
 #if _DEBUG
-                ToConsole("%s", errorMessage.c_str());
+                ToConsole(errorMessage.c_str());
 #endif
                 ToFile(errorMessage.c_str());
             }
@@ -113,15 +92,6 @@ namespace common::log {
         ToConsole("%s: %s", message, detail);
 #endif
         ToFile("%s: %s", message, detail);
-    }
-
-    void ToConsole(const char* _Format, ...) {
-        OpenConsole();
-        va_list args;
-        va_start(args, _Format);
-        vprintf(_Format, args);
-        printf("\n");
-        va_end(args);
     }
 
     void FpsToConsole() {
