@@ -1,13 +1,6 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿namespace NeoLuaBootstrap;
 
-namespace NeoLuaBootstrap
+static class Constants
 {
-    internal class Constants
-    {
-        public const string AppName = "ThMouseX";
-    }
+    public const string AppName = "ThMouseX";
 }
