@@ -39,20 +39,28 @@ namespace imguioverlay = core::imguioverlay;
 
 #define TAG "[DirectX11] "
 
-#define Vt0 XMVECTOR()
-
-#define RGBA(r, g, b, a) XMVECTORF32{(r&0xFF)/255.f, (g&0xFF)/255.f, (b&0xFF)/255.f, (a&0xFF)/255.f}
-
-#define ToneColor(i) RGBA(i, i, i, 255)
-
-constexpr auto ResizeBuffersIdx = 13;
-constexpr auto PresentIdx = 8;
-
 using namespace std;
 using namespace DirectX;
 using namespace Microsoft::WRL;
 
+static constexpr XMVECTORF32 RGBA(UCHAR r, UCHAR g, UCHAR b, UCHAR a) {
+    return XMVECTORF32{ (r & 0xFF) / 255.f, (g & 0xFF) / 255.f, (b & 0xFF) / 255.f, (a & 0xFF) / 255.f };
+}
+
+static constexpr XMVECTORF32 ToneColor(UCHAR i) {
+    return RGBA(i, i, i, 255);
+}
+
+constexpr auto ResizeBuffersIdx = 13;
+constexpr auto PresentIdx = 8;
+
 using CallbackType = void (*)(void);
+using ID3D11DevicePtr = ID3D11Device*;
+using ID3D11DeviceContextPtr = ID3D11DeviceContext*;
+using ID3D11RenderTargetViewPtr = ID3D11RenderTargetView*;
+using SpriteBatchPtr = SpriteBatch*;
+using ID3D11PixelShaderPtr = ID3D11PixelShader*;
+using ID3D11ShaderResourceViewPtr = ID3D11ShaderResourceView*;
 
 namespace core::directx11 {
     HRESULT WINAPI D3DResizeBuffers(IDXGISwapChain* swapChain, UINT BufferCount, UINT Width, UINT Height, DXGI_FORMAT NewFormat, UINT SwapChainFlags);
@@ -69,25 +77,25 @@ namespace core::directx11 {
 
     bool imGuiPrepared;
 
-    void ClearMeasurementFlags() {
+    static void ClearMeasurementFlags() {
         measurementPrepared = false;
         cursorStatePrepared = false;
     }
 
     // cursor and screen state
-    ID3D11Device*               device;
-    ID3D11DeviceContext*        context;
-    ID3D11RenderTargetView*     renderTargetView;
-    SpriteBatch*                spriteBatch;
-    ID3D11PixelShader*          pixelShader;
-    ID3D11ShaderResourceView*   cursorTexture;
+    ID3D11DevicePtr             device;
+    ID3D11DeviceContextPtr      context;
+    ID3D11RenderTargetViewPtr   renderTargetView;
+    SpriteBatchPtr              spriteBatch;
+    ID3D11PixelShaderPtr        pixelShader;
+    ID3D11ShaderResourceViewPtr cursorTexture;
     XMVECTORF32                 cursorPivot;
     XMVECTORF32                 cursorScale;
     float                       d3dScale = 1.f;
     float                       imGuiMousePosScaleX = 1.f;
     float                       imGuiMousePosScaleY = 1.f;
 
-    void CleanUp(bool forReal = false) {
+    static void CleanUp(bool forReal = false) {
         ImGui_ImplDX11_InvalidateDeviceObjects();
         SAFE_RELEASE(pixelShader);
         SAFE_DELETE(spriteBatch);
@@ -106,7 +114,7 @@ namespace core::directx11 {
         }
     }
 
-    void TearDownCallback(bool isProcessTerminating) {
+    static void TearDownCallback(bool isProcessTerminating) {
         if (isProcessTerminating)
             return;
         CleanUp(true);
@@ -132,13 +140,13 @@ namespace core::directx11 {
             return;
         }
 
-        WindowHandle tmpWnd(CreateWindowA("BUTTON", "Temp Window", WS_SYSMENU | WS_MINIMIZEBOX, CW_USEDEFAULT, CW_USEDEFAULT, 300, 300, NULL, NULL, NULL, NULL));
+        WindowHandle tmpWnd(CreateWindowA("BUTTON", "Temp Window", WS_SYSMENU | WS_MINIMIZEBOX, CW_USEDEFAULT, CW_USEDEFAULT, 300, 300, nil, nil, nil, nil));
         if (!tmpWnd) {
             note::LastErrorToFile(TAG "Failed to create a temporary window.");
             return;
         }
 
-        ComPtr<ID3D11Device> device;
+        ComPtr<ID3D11Device> _device;
         ComPtr<IDXGISwapChain> swap_chain;
         DXGI_SWAP_CHAIN_DESC sd{
             .BufferDesc = DXGI_MODE_DESC{
@@ -153,26 +161,25 @@ namespace core::directx11 {
             .Windowed = TRUE,
             .SwapEffect = DXGI_SWAP_EFFECT_DISCARD
         };
-        const D3D_FEATURE_LEVEL feature_levels[] = {D3D_FEATURE_LEVEL_11_0, D3D_FEATURE_LEVEL_10_0};
-        auto rs = _D3D11CreateDeviceAndSwapChain(NULL, D3D_DRIVER_TYPE_HARDWARE, NULL, 0, feature_levels, 2, D3D11_SDK_VERSION, &sd, &swap_chain, &device, NULL, NULL);
+        D3D_FEATURE_LEVEL feature_levels[] = { D3D_FEATURE_LEVEL_11_0, D3D_FEATURE_LEVEL_10_0 };
+        auto rs = _D3D11CreateDeviceAndSwapChain(nil, D3D_DRIVER_TYPE_HARDWARE, nil, 0, feature_levels, 2, D3D11_SDK_VERSION, &sd, &swap_chain, &_device, nil, nil);
         if (FAILED(rs)) {
             note::DxErrToFile(TAG "Failed to create device and swapchain of DirectX 11.", rs);
             return;
         }
 
-        auto baseAddress = (DWORD)d3d11;
         auto vtable = *(DWORD**)swap_chain.Get();
 
         callbackstore::RegisterUninitializeCallback(TearDownCallback);
         callbackstore::RegisterClearMeasurementFlagsCallback(ClearMeasurementFlags);
 
         minhook::CreateHook(vector<minhook::HookConfig>{
-            { PVOID(vtable[PresentIdx]), & D3DPresent, (PVOID*)&OriPresent },
-            {PVOID(vtable[ResizeBuffersIdx]), &D3DResizeBuffers, (PVOID*)&OriResizeBuffers},
+            { PVOID(vtable[PresentIdx]), & D3DPresent, &OriPresent },
+            { PVOID(vtable[ResizeBuffersIdx]), &D3DResizeBuffers, &OriResizeBuffers },
         });
     }
 
-    void PrepareFirstStep(IDXGISwapChain* swapChain) {
+    static void PrepareFirstStep(IDXGISwapChain* swapChain) {
         if (firstStepPrepared)
             return;
         firstStepPrepared = true;
@@ -189,7 +196,7 @@ namespace core::directx11 {
             note::DxErrToFile(TAG "PrepareFirstStep: swapChain->GetBuffer failed", rs);
             return;
         }
-        rs = device->CreateRenderTargetView(pBackBuffer.Get(), NULL, &renderTargetView);
+        rs = device->CreateRenderTargetView(pBackBuffer.Get(), nil, &renderTargetView);
         if (FAILED(rs)) {
             note::DxErrToFile(TAG "PrepareFirstStep: device->CreateRenderTargetView failed", rs);
             return;
@@ -199,7 +206,7 @@ namespace core::directx11 {
 
         spriteBatch = new SpriteBatch(context);
 
-        rs = device->CreatePixelShader(additiveToneShaderBlob, ARRAYSIZE(additiveToneShaderBlob), NULL, &pixelShader);
+        rs = device->CreatePixelShader(additiveToneShaderBlob, ARRAYSIZE(additiveToneShaderBlob), nil, &pixelShader);
         if (FAILED(rs)) {
             note::DxErrToFile(TAG "PrepareFirstStep: device->CreatePixelShader failed", rs);
             return;
@@ -214,14 +221,14 @@ namespace core::directx11 {
         g_hFocusWindow = desc.OutputWindow;
         g_isMinimized = IsIconic(g_hFocusWindow);
 
-        if (gs_textureFilePath[0] && SUCCEEDED(CreateWICTextureFromFile(device, gs_textureFilePath, NULL, &cursorTexture))) {
+        if (gs_textureFilePath[0] && SUCCEEDED(CreateWICTextureFromFile(device, gs_textureFilePath, nil, &cursorTexture))) {
             ComPtr<ID3D11Resource> resource;
             cursorTexture->GetResource(&resource);
             ComPtr<ID3D11Texture2D> pTextureInterface;
             resource->QueryInterface<ID3D11Texture2D>(&pTextureInterface);
-            D3D11_TEXTURE2D_DESC desc;
-            pTextureInterface->GetDesc(&desc);
-            cursorPivot = {(desc.Height - 1) / 2.f, (desc.Width - 1) / 2.f, 0.f};
+            D3D11_TEXTURE2D_DESC _desc;
+            pTextureInterface->GetDesc(&_desc);
+            cursorPivot = { (_desc.Height - 1) / 2.f, (_desc.Width - 1) / 2.f, 0.f };
         }
     }
 
@@ -236,7 +243,7 @@ namespace core::directx11 {
     - determine g_pixelRate
     - determine g_pixelOffset
     */
-    void PrepareMeasurement(IDXGISwapChain* swapChain) {
+    static void PrepareMeasurement(IDXGISwapChain* swapChain) {
         if (measurementPrepared)
             return;
         measurementPrepared = true;
@@ -264,17 +271,17 @@ namespace core::directx11 {
             note::LastErrorToFile(TAG "PrepareMeasurement: GetClientRect failed");
             return;
         }
-        g_pixelRate = float(g_currentConfig.BaseHeight) / clientSize.height();
+        g_pixelRate = float(g_currentConfig.BaseHeight) / float(clientSize.height());
         g_pixelOffset.X = g_currentConfig.BasePixelOffset.X / g_pixelRate;
         g_pixelOffset.Y = g_currentConfig.BasePixelOffset.Y / g_pixelRate;
-        imGuiMousePosScaleX = float(clientSize.width()) / desc.BufferDesc.Width;
-        imGuiMousePosScaleY = float(clientSize.height()) / desc.BufferDesc.Height;
+        imGuiMousePosScaleX = float(clientSize.width()) / float(desc.BufferDesc.Width);
+        imGuiMousePosScaleY = float(clientSize.height()) / float(desc.BufferDesc.Height);
     }
 
     /*
     Determine scaling
     */
-    void PrepareCursorState(IDXGISwapChain* swapChain) {
+    static void PrepareCursorState(IDXGISwapChain* swapChain) {
         if (cursorStatePrepared)
             return;
         cursorStatePrepared = true;
@@ -289,40 +296,41 @@ namespace core::directx11 {
             return;
         }
 
-        auto scale = float(desc.BufferDesc.Height) / gs_textureBaseHeight;
-        cursorScale = XMVECTORF32{scale, scale};
+        auto scale = float(desc.BufferDesc.Height) / float(gs_textureBaseHeight);
+        cursorScale = XMVECTORF32{ scale, scale };
 
         RECTSIZE clientSize{};
         if (GetClientRect(g_hFocusWindow, &clientSize) == FALSE) {
             note::LastErrorToFile(TAG "PrepareCursorState: GetClientRect failed");
             return;
         }
-        d3dScale = float(clientSize.width()) / desc.BufferDesc.Width;
+        d3dScale = float(clientSize.width()) / float(desc.BufferDesc.Width);
     }
 
-    void RenderCursor(IDXGISwapChain* swapChain) {
+    static void RenderCursor(IDXGISwapChain* swapChain) {
+        using enum ModulateStage;
         if (!cursorTexture || !spriteBatch || !renderTargetView || !device || !context)
             return;
-                
+
         auto dx11State = graphics::SaveDx11State(swapChain, context);
 
         // scale mouse cursor's position from screen coordinate to D3D coordinate
         auto pointerPosition = helper::GetPointerPosition();
-        XMVECTOR cursorPositionD3D = XMVECTORF32{float(pointerPosition.x), float(pointerPosition.y)};
+        XMVECTOR cursorPositionD3D = XMVECTORF32{ float(pointerPosition.x), float(pointerPosition.y) };
         if (d3dScale != 0.f && d3dScale != 1.f) {
             cursorPositionD3D = XMVectorScale(cursorPositionD3D, d3dScale);
         }
 
         // scale cursor sprite to match the current render resolution
-        auto scalingMatrixD3D = XMMatrixTransformation2D(cursorPositionD3D, 0, cursorScale, Vt0, 0, Vt0);
+        auto scalingMatrixD3D = XMMatrixTransformation2D(cursorPositionD3D, 0, cursorScale, XMVECTOR(), 0, XMVECTOR());
 
-        context->OMSetRenderTargets(1, &renderTargetView, NULL);
+        context->OMSetRenderTargets(1, &renderTargetView, nil);
 
         static UCHAR tone = 0;
         static auto toneStage = WhiteInc;
         auto usePixelShader = false;
         if (g_inputEnabled) {
-            helper::CalculateNextTone(_ref tone, _ref toneStage);
+            helper::CalculateNextTone(tone, toneStage);
             if (toneStage == WhiteInc || toneStage == WhiteDec)
                 // default behaviour: texture color * diffuse color
                 // this shader: texture color + diffuse color (except alpha)
@@ -330,18 +338,18 @@ namespace core::directx11 {
         }
 
         // draw the cursor
-        auto setCustomShaders = usePixelShader ? []() { context->PSSetShader(pixelShader, NULL, 0); } : nullptr;
+        auto setCustomShaders = usePixelShader ? []() { context->PSSetShader(pixelShader, nil, 0); } : nil;
         auto sortMode = SpriteSortMode_Deferred;
         auto m_states = std::make_unique<CommonStates>(device);
-        spriteBatch->Begin(sortMode, m_states->NonPremultiplied(), NULL, NULL, NULL, setCustomShaders, scalingMatrixD3D);
+        spriteBatch->Begin(sortMode, m_states->NonPremultiplied(), nil, nil, nil, setCustomShaders, scalingMatrixD3D);
         auto color = g_inputEnabled ? ToneColor(tone) : RGBA(255, 200, 200, 128);
-        spriteBatch->Draw(cursorTexture, cursorPositionD3D, NULL, color, 0, cursorPivot, 1, SpriteEffects_None);
+        spriteBatch->Draw(cursorTexture, cursorPositionD3D, nil, color, 0, cursorPivot, 1, SpriteEffects_None);
         spriteBatch->End();
-        
+
         graphics::LoadDx11State(context, dx11State);
     }
 
-    void PrepareImGui() {
+    static void PrepareImGui() {
         if (imGuiPrepared)
             return;
         imGuiPrepared = true;
@@ -353,7 +361,7 @@ namespace core::directx11 {
         ImGui_ImplDX11_Init(device, context);
     }
 
-    void ConfigureImGui(IDXGISwapChain* swapChain) {
+    static void ConfigureImGui(IDXGISwapChain* swapChain) {
         if (imGuiConfigured)
             return;
         imGuiConfigured = true;
@@ -368,10 +376,10 @@ namespace core::directx11 {
             return;
         }
 
-        imguioverlay::Configure(float(desc.BufferDesc.Height) / gs_imGuiBaseVerticalResolution);
+        imguioverlay::Configure(float(desc.BufferDesc.Height) / float(gs_imGuiBaseVerticalResolution));
     }
 
-    void RenderImGui(IDXGISwapChain* swapChain) {
+    static void RenderImGui(IDXGISwapChain* swapChain) {
         if (!g_showImGui || !context || !renderTargetView)
             return;
         ImGui_ImplDX11_NewFrame();
@@ -386,7 +394,7 @@ namespace core::directx11 {
             desc.BufferDesc.Width, desc.BufferDesc.Height,
             imGuiMousePosScaleX, imGuiMousePosScaleY
         );
-        context->OMSetRenderTargets(1, &renderTargetView, NULL);
+        context->OMSetRenderTargets(1, &renderTargetView, nil);
         ImGui_ImplDX11_RenderDrawData(drawData);
     }
 
