@@ -32,12 +32,15 @@ namespace imguioverlay = core::imguioverlay;
 
 #define TAG "[DirectX8] "
 
-#define ToneColor(i) D3DCOLOR_RGBA(i, i, i, 255)
+static constexpr D3DCOLOR ToneColor(UCHAR i) {
+    return D3DCOLOR_RGBA(i, i, i, 255);
+}
 
-#define SetTextureColorStage(dev, i, op, arg1, arg2)      \
-    dev->SetTextureStageState(i, D3DTSS_COLOROP, op);     \
-    dev->SetTextureStageState(i, D3DTSS_COLORARG1, arg1); \
-    dev->SetTextureStageState(i, D3DTSS_COLORARG2, arg2)
+static void SetTextureColorStage(IDirect3DDevice8* dev, DWORD stage, DWORD op, DWORD arg1, DWORD arg2) {
+    dev->SetTextureStageState(stage, D3DTSS_COLOROP, op);
+    dev->SetTextureStageState(stage, D3DTSS_COLORARG1, arg1);
+    dev->SetTextureStageState(stage, D3DTSS_COLORARG2, arg2);
+}
 
 constexpr auto CreateDeviceIdx = 15;
 
@@ -67,7 +70,7 @@ namespace core::directx8 {
 
     bool imGuiPrepared;
 
-    void ClearMeasurementFlags() {
+    static void ClearMeasurementFlags() {
         measurementPrepared = false;
         cursorStatePrepared = false;
     }
@@ -81,7 +84,7 @@ namespace core::directx8 {
     float               imGuiMousePosScaleX = 1.f;
     float               imGuiMousePosScaleY = 1.f;
 
-    void CleanUp(bool forReal = false) {
+    static void CleanUp(bool forReal = false) {
         ImGui_ImplDX8_InvalidateDeviceObjects();
         SAFE_RELEASE(cursorSprite);
         SAFE_RELEASE(cursorTexture);
@@ -102,7 +105,7 @@ namespace core::directx8 {
             pD3D, Adapter, DeviceType, hFocusWindow, BehaviorFlags, pPresentationParameters, ppReturnedDeviceInterface);
     }
 
-    void TearDownCallback(bool isProcessTerminating) {
+    static void TearDownCallback(bool isProcessTerminating) {
         if (isProcessTerminating)
             return;
         CleanUp(true);
@@ -129,7 +132,7 @@ namespace core::directx8 {
         }
 
         WindowHandle tmpWnd(CreateWindowA("BUTTON", "Temp Window",
-            WS_SYSMENU | WS_MINIMIZEBOX, CW_USEDEFAULT, CW_USEDEFAULT, 300, 300, NULL, NULL, NULL, NULL));
+            WS_SYSMENU | WS_MINIMIZEBOX, CW_USEDEFAULT, CW_USEDEFAULT, 300, 300, nil, nil, nil, nil));
         if (!tmpWnd) {
             note::LastErrorToFile(TAG "Failed to create a temporary window.");
             return;
@@ -158,7 +161,6 @@ namespace core::directx8 {
             return;
         }
 
-        auto baseAddress = (DWORD)d3d8;
         auto vtable = *(DWORD**)pD3D.Get();
         auto vtable2 = *(DWORD**)pDevice.Get();
 
@@ -166,13 +168,13 @@ namespace core::directx8 {
         callbackstore::RegisterClearMeasurementFlagsCallback(ClearMeasurementFlags);
 
         minhook::CreateHook(vector<minhook::HookConfig>{
-            { PVOID(vtable[CreateDeviceIdx]), & D3DCreateDevice, (PVOID*)&OriCreateDevice },
-            {PVOID(vtable2[ResetIdx]), &D3DReset, (PVOID*)&OriReset},
-            {PVOID(vtable2[PresentIdx]), &D3DPresent, (PVOID*)&OriPresent},
+            { PVOID(vtable[CreateDeviceIdx]), &D3DCreateDevice, &OriCreateDevice },
+            { PVOID(vtable2[ResetIdx]), &D3DReset, &OriReset },
+            { PVOID(vtable2[PresentIdx]), &D3DPresent, &OriPresent },
         });
     }
 
-    void PrepareFirstStep(IDirect3DDevice8* device) {
+    static void PrepareFirstStep(IDirect3DDevice8* device) {
         if (firstStepPrepared)
             return;
         firstStepPrepared = true;
@@ -190,7 +192,7 @@ namespace core::directx8 {
             D3DXCreateSprite(device, &cursorSprite);
             D3DSURFACE_DESC cursorSize;
             cursorTexture->GetLevelDesc(0, &cursorSize);
-            cursorPivot = {(cursorSize.Height - 1) / 2.f, (cursorSize.Width - 1) / 2.f};
+            cursorPivot = { float(cursorSize.Height - 1) / 2.f, float(cursorSize.Width - 1) / 2.f };
         }
     }
 
@@ -205,7 +207,7 @@ namespace core::directx8 {
     - determine g_pixelRate
     - determine g_pixelOffset
     */
-    void PrepareMeasurement(IDirect3DDevice8* pDevice) {
+    static void PrepareMeasurement(IDirect3DDevice8* pDevice) {
         if (measurementPrepared)
             return;
         measurementPrepared = true;
@@ -242,17 +244,17 @@ namespace core::directx8 {
             note::LastErrorToFile(TAG "PrepareMeasurement: GetClientRect failed (2)");
             return;
         }
-        g_pixelRate = float(g_currentConfig.BaseHeight) / clientSize.height();
+        g_pixelRate = float(g_currentConfig.BaseHeight) / float(clientSize.height());
         g_pixelOffset.X = g_currentConfig.BasePixelOffset.X / g_pixelRate;
         g_pixelOffset.Y = g_currentConfig.BasePixelOffset.Y / g_pixelRate;
-        imGuiMousePosScaleX = float(clientSize.width()) / d3dSize.Width;
-        imGuiMousePosScaleY = float(clientSize.height()) / d3dSize.Height;
+        imGuiMousePosScaleX = float(clientSize.width()) / float(d3dSize.Width);
+        imGuiMousePosScaleY = float(clientSize.height()) / float(d3dSize.Height);
     }
 
     /*
     Determine scaling
     */
-    void PrepareCursorState(IDirect3DDevice8* pDevice) {
+    static void PrepareCursorState(IDirect3DDevice8* pDevice) {
         if (cursorStatePrepared)
             return;
         cursorStatePrepared = true;
@@ -274,7 +276,7 @@ namespace core::directx8 {
             note::DxErrToFile(TAG "PrepareCursorState: pSurface->GetDesc failed", rs);
             return;
         }
-        auto scale = float(d3dSize.Height) / gs_textureBaseHeight;
+        auto scale = float(d3dSize.Height) / float(gs_textureBaseHeight);
         cursorScale = D3DXVECTOR2(scale, scale);
 
         RECTSIZE clientSize{};
@@ -282,10 +284,11 @@ namespace core::directx8 {
             note::LastErrorToFile(TAG "PrepareCursorState: GetClientRect failed");
             return;
         }
-        d3dScale = float(clientSize.width()) / d3dSize.Width;
+        d3dScale = float(clientSize.width()) / float(d3dSize.Width);
     }
 
-    void RenderCursor(IDirect3DDevice8* pDevice) {
+    static void RenderCursor(IDirect3DDevice8* pDevice) {
+        using enum ModulateStage;
         if (!cursorTexture)
             return;
 
@@ -322,19 +325,19 @@ namespace core::directx8 {
         if (g_inputEnabled) {
             static UCHAR tone = 0;
             static auto toneStage = WhiteInc;
-            helper::CalculateNextTone(_ref tone, _ref toneStage);
+            helper::CalculateNextTone(tone, toneStage);
             if (toneStage == WhiteInc || toneStage == WhiteDec) {
                 // default behaviour: texture color * diffuse color
                 // this:              texture color + diffuse color (except alpha)
                 SetTextureColorStage(pDevice, 0, D3DTOP_ADD, D3DTA_TEXTURE, D3DTA_DIFFUSE);
-                cursorSprite->Draw(cursorTexture, NULL, &cursorScale, NULL, 0, &cursorPositionD3D, ToneColor(tone));
+                cursorSprite->Draw(cursorTexture, nil, &cursorScale, nil, 0, &cursorPositionD3D, ToneColor(tone));
             }
             else {
-                cursorSprite->Draw(cursorTexture, NULL, &cursorScale, NULL, 0, &cursorPositionD3D, ToneColor(tone));
+                cursorSprite->Draw(cursorTexture, nil, &cursorScale, nil, 0, &cursorPositionD3D, ToneColor(tone));
             }
         }
         else {
-            cursorSprite->Draw(cursorTexture, NULL, &cursorScale, NULL, 0, &cursorPositionD3D, D3DCOLOR_RGBA(255, 200, 200, 128));
+            cursorSprite->Draw(cursorTexture, nil, &cursorScale, nil, 0, &cursorPositionD3D, D3DCOLOR_RGBA(255, 200, 200, 128));
         }
         cursorSprite->End();
 
@@ -344,7 +347,7 @@ namespace core::directx8 {
         pDevice->EndScene();
     }
 
-    void PrepareImGui(IDirect3DDevice8* pDevice) {
+    static void PrepareImGui(IDirect3DDevice8* pDevice) {
         if (imGuiPrepared)
             return;
         imGuiPrepared = true;
@@ -357,7 +360,7 @@ namespace core::directx8 {
         ImGui_ImplDX8_Init(pDevice);
     }
 
-    void ConfigureImGui(IDirect3DDevice8* pDevice) {
+    static void ConfigureImGui(IDirect3DDevice8* pDevice) {
         if (imGuiConfigured)
             return;
         imGuiConfigured = true;
@@ -378,10 +381,10 @@ namespace core::directx8 {
             return;
         }
 
-        imguioverlay::Configure(float(d3dSize.Height) / gs_imGuiBaseVerticalResolution);
+        imguioverlay::Configure(float(d3dSize.Height) / float(gs_imGuiBaseVerticalResolution));
     }
 
-    void RenderImGui(IDirect3DDevice8* pDevice) {
+    static void RenderImGui(IDirect3DDevice8* pDevice) {
         if (!g_showImGui)
             return;
         ImGui_ImplDX8_NewFrame();
